LoW: delete brackets from "wmllint: who" comments
The brackets are a relic from when I had a different conception of how the
"who ... is ..." comment would work. Although these comments still work,
because I now strip brackets from the macro name, I don't want UMC authors
getting the idea that this is the "right" way to do these comments.


Former-commit-id: 5c080ed41dbfef22b8747d66a86a62efb04d7297
Former-commit-id: 9665d77c987b6c1e133df221aa604bef8c701998
Former-commit-id: f2535b44e82062c0602f6423202a319fdf600c78
Former-commit-id: 663ee6615efc72bc923005abfa18e8d9c882b555 [formerly 5a190613957606a3e02c9a635b7c6449fbf9d795]
Former-commit-id: d51ad1cf900a9ce89e0e855c99c7023e7fb9a569