NR 05a: change "Should this unit" to "Should I" in Rod dialog
The other choice dialogs in this scenario use "I" or "we", so
this brings it into conformity with them.

On the mailing list, I also spoke of a desire to have Camerin's
lines run before the choice is made, but looking at it, this
is a more complex change than I want to deal with right now.


Former-commit-id: bbde90c1cb700ba1e7e732f8c7ee4b397d88a1b1
Former-commit-id: 57a8a7c4149c540d7f7e04085142a77e6b2c39d1
Former-commit-id: 23db01d0ddcbd63e35b290f8c9ddc4a6799464aa
Former-commit-id: 9e9859a847ba880f26e6dd39d31ad85bf9ddcd59 [formerly b3cf281fda85f67476e060939dd8533ebe2d2ae3]
Former-commit-id: b3006713154e46ceb145f6ec232b5b8744c64e02