campaign: Implement [server_info] block and feedback_url_format attribute
The [server_info] configuration block is intended to contain additional
information about a campaignd instance that may be used by clients or
the server for various purposes.

As a first such information field, the feedback_url_format string
contains a format string for use with
fast_interpolate_variables_into_string() to build add-on feedback page
URLs.


Former-commit-id: 567df4accc803748752822e5a8ea754697672996
Former-commit-id: f2b56da50d9b0aaae9d3ac24ab22993e9c20b304
Former-commit-id: 4d8e2733eec86faf63a332898a631f9bc171df75
Former-commit-id: 1bd2a9016429b14e16b33dbfd46a15e08a3ecb07 [formerly 6b334e91fd321f8bec572965374c9ac0de24491d]
Former-commit-id: 21307923506dd07246d18dec1238e72a30c5db8b