gui2/taddon_description: Completely redesigned dialog layout
The new layout merges some information rows together to increase
available vertical space for displaying additional fields later, and
uses a 'description' scroll_label type for the add-on Description field
to improve its appearance, especially for multi-line descriptions.

This also makes the dialog non-click-dismissable and adds a Close button
on the bottom right corner. Since we're about to add more interactive
controls, there's really not much of a choice about the click-dismiss
option; it gets disabled for such dialogs automatically.


Former-commit-id: f5b73949b5e17f1e90086c567d5b3e3388efc652
Former-commit-id: 88b0a2ccb9ec2c47f42ab5141905b83081b7aa02
Former-commit-id: 05782894f029a887f14c143cf471bca719f2b3bd
Former-commit-id: 94fab0d206905d3d4c0ee54ea1a161e1acbeafd6 [formerly eac7cafffeb845d01d9e1889f185ccb826363f1a]
Former-commit-id: 7cb78f21e1e90360a9cab2253067d4aca94f3724