Implementation of a "smart" linked list.
With a standard linked list, iterators are never invalidated unless
they point to an element that was erased (or the list is destroyed).
This implementation takes that one step further and never invalidates
iterators (unless the list is destroyed). From the perspective of the
list, erased elements no longer exist, but an iterator that still
points to an erased element can be dereferenced, incremented, and
decremented.

I was having trouble implementing this based on STL and Boost, so
I went ahead and coded this from scratch. The standard list interface
is implemented, aside from assignment (since I could not reconcile
assignment with the goal of not invalidating iterators).


Former-commit-id: baf65fea882fb499818c50a4b7a52c78f6a4a2c4
Former-commit-id: a1c478cdcc8f5bdfb5b3debbc7cb60e253e23d85
Former-commit-id: 8388ddd441a00d28079d1eac9495324f760d55f9
Former-commit-id: cdb4c1ad71fd759f2606f7dfb85a63dc6259239f [formerly ed23444f3ca9ec7b1e94d1a7a0493f266926ee9c]
Former-commit-id: abd01a574c65b555cd7e483a25b4800bd1e3d116