SotBE 04: fix ORCISH_SHAMAN id recognition
While I tested my whofield feature, unfortunately it was on an
old version of the campaign in which all shamans were on side 1,
so the macros had no side field. Recently, it was decided that
the shamans in this scenario should start on side 3, and a side
field was inserted into the macros.

When it was time to commit, there was of course a conflict, but
I simply put my existing magic comments into the new version,
not considering that now the id field had been shifted to the
right by one, and didn't re-run wmllint.


Former-commit-id: 43a2462295602fc93a89b5866bf8603305575ba4
Former-commit-id: 6a8fc003c6e35b17701a991cdd661f6433df5ce6
Former-commit-id: 8248c64b855c95d5c75a1e9c168be109fc12438a
Former-commit-id: 6100ace434c1de5f14d69a037808b88d47e6b64b [formerly 51b540a5d53df448c3d43b215e72983c58726885]
Former-commit-id: 8ec7522264f45166c3d6d7ee1f6dd2a7bf0587b1