SotBE 17: recognize speaker Bruce
While I looked for "unknown speakers" in an old version of the campaign,
in the meantime mattsc had changed the scenario to use a macro for the
enemy reinforcements (e5905f05).

Although I could use my spiffy new "whofield" magic comment here, Bruce
is the only general who speaks, so an old-fashioned "recognize" is more
appropriate.


Former-commit-id: f97122f4ac7d734f2cacbea278a84293dce3517c
Former-commit-id: ed82a52d245e366a33d9f867f17020841f82632e
Former-commit-id: 6a624b00a7f6069af3fae7e8afe2bdba41252961
Former-commit-id: 7f735895014b760cd1143efd62320e271519880f [formerly e7a16437309235300b1563d735a6dd718797faba]
Former-commit-id: 87b611cae60029c4de079b6c0df700096ff09970