New wesnoth-defaults utility script
I wrote this a long time ago and have been using it to ease testing of
Wesnoth's default configuration under different conditions ever since.

It creates a new configuration and data directory under /tmp that only
exists for the duration of the Wesnoth session, and disables WML cache
functionality.

This is extremely useful for quickly testing and debugging functionality
dependant on user configuration settings.


Former-commit-id: 78df08010b765673aeefa04a7eb7e68b6e41916c
Former-commit-id: 45a83b4c3589336b96d89a93615817cf12d143be
Former-commit-id: b41f239b4c04c2fc8b7473a933dbfb2418cded0a
Former-commit-id: 9b8d02307ce0da7ee82971606e51a0b99fbaf8a5 [formerly 61bd0343099d120fae511a5255a06f923e4d48ed]
Former-commit-id: 2b161f58e09c99bbd6c5ccf03f7d4ee639133adc