Truncate long Advanced Preferences entries with ellipses
This truncates long Advanced Preferences list entry labels and values
with ellipses using hardcoded character count limits in order to avoid
stretching out the listbox too much over the dialog's width.

Currently, this is notably necessary for the German translation.

The character count limits chosen for this commit (46 for the label, 8
for the value) only really work with the basic ASCII character set. Some
Unicode characters may be rendered wider. Furthermore, the Preferences
page selection list may container longer entries that push this listbox
further to the right, closer to the dialog's right edge.

Also, I really hope most languages have words for 'yes' or 'no' that are
at most 8 characters in length or can at least be read without ambiguity
when truncated to that limit. At least none of the translations we
support which have translated those words hit the limit.


Former-commit-id: 8aecf60e10b2cd7b0b3b6ec9f6085417c51cb182
Former-commit-id: 1b2a641b9723e612d3f3833e6a0b028b144e153e
Former-commit-id: 87fbdfcf248ffa32fba442fbb8483bbcc4a3293d
Former-commit-id: f008eef60620ccd5d7ee74c5d747b9f2d41225e6 [formerly 155c935e8daeea53c51283f9cf351e83225ad6ac]
Former-commit-id: 6bdef6d764563001cb45323fe89374f032f6f11c