campaignd: Don't prefix errors sent to clients with "#Error: "
The string only worked as intended before the GUI2/Pango days (i.e.
before some 1.5.x version), meaning that the client would use a red
color for the font.

In other words, deplorable user interface design.

Nowadays, the add-ons client uses a dedicated error message path for
presenting this kind of alerts to the user, which uses a dialog with a
visible "Error" caption and everything. We don't really need extra
formatting, and in any case, *it should not be the server's mission to
do this*!


Former-commit-id: 88f993b0198c4f2c85915f482a8e33dd6091b696
Former-commit-id: 0ce2460c81d52f26beb28ff540bd056888badd3b
Former-commit-id: 4ec96a6ec00c9e9ca81741636f6f9f167848840b
Former-commit-id: 6db6169c4674ac42f4df6c8d882e99e45293990a [formerly a38ea399997362dc30aee42d9fa3a0ea45ad0ec1]
Former-commit-id: 2950c072a4eebd76f65e0e1d60b6ff14bf35bbc8