Fix logic of [filter_vision] in Standard Unit Filters (SUFs)
Previously, the [filter_vision] tag would cause a false positive when a
unit was on terrain on which it was invisible and that also was under
fog/shroud.  There might have been other combinations that did not work
either, I did not check that.  I did check that all 8 combinations of
visble=yes/no, the unit hiding due to terrain or not, and the hex being
under fog/shroud or not work now.


Former-commit-id: eca1c28d3c0100b9376d12dce5b32ef81ad1b003
Former-commit-id: 7f0c7e9b74d264a4b1a912ae3ad315826086a2ff
Former-commit-id: 341de90aa10f179fd498cdce2afa3c434ba0d18a
Former-commit-id: 31ce8e0f2cc91dc4ee54d3a8ac1e50dfb406b43c [formerly 3859a093be729bb288ccb8878774e22b09c1e962]
Former-commit-id: 283f58f4f1ccaa83b50001d3eddca2060e9b1fe9