wmllint: report stored units never unstored or cleared
Keys from the dictionary of stored units are removed as the unit is
unstored or its variable cleared. However, I found that one character in
Legend of Wesmere, Urudin, is stored but apparently never unstored/cleared.
I figure it's best to report such cases.

This message does not include helpful information like the filename,
because that data was never stored in the dictionary. The dictionary was
designed on the presumption that all entries would have a matching unstore/
clear event, and I didn't think of this warning until I saw there was an
anomalous case. Those who get this error will just have to use grep!


Former-commit-id: b34b4a1010ee5812831fb20e279757dabf39dd9b
Former-commit-id: 8a51445c04f7814080a8d8aa1a7bee9533f59391
Former-commit-id: eae7d9b5d817d896ef6ac00c21f25e4221d15a56
Former-commit-id: f240dd4d84666f35485fd93da77bc89a0d9776da [formerly 1ca84af9f90442c7656df58acb864ef11236a428]
Former-commit-id: 7261a4b901294275c2ee2fb73abf08be9b9a36fd