Add missing check for a null surface from image::get_image()
Fixes a NULL pointer dereference when drawing preview pane overlays in
the Recall Unit dialog for a unit whose overlays list includes an image
that doesn't exist or can't be read, and debug mode is not enabled.

This didn't happen in debug mode because image::get_image() retrieves a
valid placeholder surface (currently images/misc/missing-image.png) in
that situation with debug mode enabled.

Reported by SkyOne in the forums: http://r.wesnoth.org/t39983

Thanks to mattsc for pinpointing the commit at fault.

Note: the code handling overlays on the listbox/menu of the Recall Unit
dialog shouldn't be affected by this since it uses the BLIT image path
function instead of accessing surfaces directly.

Former-commit-id: bec14393d701899cd4ac1c5654756409cb9e4415
Former-commit-id: 3e02c64e26926c745f388e98f3231c206f3cc1ad
Former-commit-id: 1d6e411b2a91c0d32fe6a5f53da5d6fa34b6c46d
Former-commit-id: a8ea1e5f0233dc6243ba49f0660813cc810ae6b1 [formerly a2cd7c328af2e56a46b3c186fdd58350f7750f36]
Former-commit-id: 28edf71b49ec4b17ee48206c1aeeb81750e53873