gui2/twml_error: New GUI2 dialog to be used for WML parser/pp errors
This dialog features separate summary and detail sections along with a
small heading for the latter in bold face. More features may be added
later as part of my work on improving WML diagnostics, but for now we
have the bare minimum for better message formatting with a content
structure more or less identical to the current one.


Former-commit-id: 7ebe507646e8223f98c299ea33fad8e49c2c9249
Former-commit-id: a4be16b806a60462123996161789b1953708ed0e
Former-commit-id: 04174d9a0ef1ba193d4e23f3deca7ab5ce459a1b
Former-commit-id: 7d25bbd6534f832ce83bbe698833c073c7e1e45d [formerly 0d2ec93321e428aa1db52afb21294128551f1232]
Former-commit-id: 61866f2501856a8af617b81e927741243c8482f0