gui2/twml_error: Handle an optional list of faulty files separately
This allows us to be more self-contained and require less logic in the
instantiation site for preparing the report for display.

This requires adding a new row and label for displaying the list of
faulty files, which will be hidden when the list is empty or not
provided in the display() static member function.


Former-commit-id: ceb1fd144e9df945b73292988d4c066204c1ccab
Former-commit-id: abe2cf95f18f08126c87c162073b9e0f72d5223f
Former-commit-id: 0287c826e47f75d73015b92954065b134ef5e4da
Former-commit-id: 5266ce1d5ca77befb887c199927e4ffc199d4b3f [formerly b1ff5fe2cdc1441c299ff70f85deac02b02deac5]
Former-commit-id: a99a3da49eedd3ded7942e8f0ef35e569327f323