preprocessor: Indent error location info on a separate line
Messages such as:

> Quoted string not terminated at ~add-ons/Foo/bar.cfg
> Macro/file 'FOOBAR' is missing at ~add-ons/Baz/bat.cfg:12 included from ~add-ons/Baz/bag.cfg:2 included from ~add-ons/Baz/ban.cfg:42

Now are formatted as:

> Quoted string not terminated
>     at ~add-ons/Foo/bar.cfg
> Macro/file 'FOOBAR' is missing
>    at ~add-ons/Baz/bat.cfg:12
>    included from ~add-ons/Baz/bag.cfg:2
>    included from ~add-ons/Baz/ban.cfg:42

For messages that include the {} substitution trail, this improves
readability by making the trail a sequence of lines with file locations
rather than a massive opaque and unreadable text wall:

> Macro/file 'FOOBAR' is missing at ~add-ons/After_the_Storm//macros/scenario-segment-implementation.cfg:61 included from ~add-ons/After_the_Storm/base-loader.cfg:6 included from ~add-ons/After_the_Storm/_main.cfg:174 included from ~add-ons/After_the_Storm/base-loader.cfg:75 included from [...]

In fact, it's that particular case that motivated this report format
change. Cases where there is no substitution trail but there is still a
file location pointer are covered by this commit too mainly for the sake
of consistency and readability -- being able to see the last (first)
relevant location pointer first on a separate line should make it easier
to debug most WML issues.


Former-commit-id: e45cb05b72d24b7667f47f57ce82b1a0343a2e92
Former-commit-id: ae3cb11bfe8c103467e5999a53758342d84e042f
Former-commit-id: a422e80ceb0342c605d0ba2914f47b36051cefb2
Former-commit-id: 810d3eb1a669570d67b3353dd7dff8998a9369d5 [formerly b221106aac364f57b1df22d1ad8ebaf247fd32c2]
Former-commit-id: b2e6fd92008beba4ab14f2ab3842e9bbd52d2d6d