gui2/tchat_log: Hack to set a minimum cell size to 200x250
Because this window many contain a lot of text in both directions,
having it change size dynamically every time the log page is switched or
a new text filter is entered can be extremely annoying for the player.

Setting the scroll_label's parent (now ancestor) cell's minimum size to
something that isn't 0x0 mitigates the issue to some extent. Better than
nothing, I guess.

(Tested with a screen resolution of 800x480.)


Former-commit-id: 62279d28b144253f5648d6041af63947d4f71744
Former-commit-id: 2e5ad6dc51c326e7a7a2dafc5a991f1fcebf4619
Former-commit-id: 5240db41ee7c7277d53d2f91a275501db82cb1ff
Former-commit-id: 731495f9df89e3644675af211149879a1bcc6566 [formerly 52847cf71ad79f93413fc1aa4da3f423ff16983a]
Former-commit-id: 77cc6713b81cb7c2deb6c6f33de84bbafa23a8a7