Fix drawing the offmap area for small resolutions.
The problem was a bit tricky. A partial resolution removed an element with
an empty id. The main_map_border section has no id and thus an empty id.
So the inheritance removed this section by accident. Of course the id for
a removal or change should be mandatory, which the next patch will do.

This fixes a part of bug #21584.



Former-commit-id: 0ef0f62bba053838fc2d35c8491856ff2f3df920
Former-commit-id: e88bb72e162e6b78129ba63364546eb36cfd4976
Former-commit-id: 780e8c3c7dacab3ba3f6689775680e0f8f10cbec
Former-commit-id: 6f76420d6336a84628592dc15801932dfa21a3d2 [formerly f1b45cb97e5897875dcc7d385416964de1d50dd5]
Former-commit-id: e4f0b2040f2ab6f072bf655f42046315dc892950