Disallow change and remove sections without an id.
This change to the ThemeWML will avoid the offmap drawing issues due to
accidentaly changing or removing the main_map_border section from a
resolution. If it is really wanted it is possible to add an id to this
section, the game doesn't use it, but inheritance uses the raw config.

This fix rewrites the expand_partialresolution, adding the check and
making the code easier to follow.

This should avoid bugs like bug #21584.


Former-commit-id: 394e7ae4f6429a5923fa9cc2108f046344c89fec
Former-commit-id: e9a4b450da21f0f5cc4c257056f0bc99ed1760e8
Former-commit-id: c63f9ecfc3e2d795f826928e67d9a8d45417657d
Former-commit-id: d7cc42dff69c88e379f8f28862891e17eee7139d [formerly 3f4a301afb90300fc7cd37d7cf5827ce0b9ae4ab]
Former-commit-id: fc5a001c160788981593adebc35a8b34d694739d