Add a work-around for a blurring issue.
When pressing the next or previous tip button there is a drawing glitch.
As soon as the mouse pointer leaves the button area the glitch
disappears. For now force an update directly after the button is
pressed. This is not particularly elegant, but it works. The SDL2
transition will probably fix this issue automatically.


Former-commit-id: 90e37e5a0277b5f6992f29528aa2c0fd8c636e29
Former-commit-id: c3923aeaf912a1f240acfa495e0e574f4c217387
Former-commit-id: 7565832a8a3c461e393dfbdb4b177d9d4616ab14
Former-commit-id: 80c949f4a9a2289319ae18c1722916114851d4ef [formerly 8511070c93973f7cf285b121f98e913635c9dc70]
Former-commit-id: d3c5258577f4ed2ac495317eb17baa9e9aaef6d9