Optimizing game event firing.
This uses the indexing of handlers by name to reduce the number of
event handlers that are checked when an event is fired. There is
extra overhead to independently track handlers whose names contain
variables, but in more common cases, there should be an improvement,
particularly when many events fire in sequence, such as the enter/
exit_hex events.


Former-commit-id: 42850671211de16561fe6e6bb1eb48fa5ed86dd9
Former-commit-id: 34d0a563d555b575354d1fb9ed1090ad92ab083d
Former-commit-id: fc916b2b88fb7982f084ba269d4f34486f77d983
Former-commit-id: 40b571c3f9c6bf26e2a1f27cecc354f3e16aeaae [formerly bf076d01edbe67692c59dcf9ca1e4f33c94971f4]
Former-commit-id: 2729d8981e728d7efba5354757e17f0170dc0d42