theme ui, gui2/ttheme_list: New theme chooser, implement [theme] description
Adds a new 'description' attribute to [theme] and replaces the previous
instance of gui2::tsimple_item_selector used to select themes with a
brand new dialog (that's actually based on it) that allows for
displaying theme descriptions in a nice format.

Because some of the theme names aren't obvious enough to players
("UnitBox" comes to mind), we ought to provide descriptions of each
theme to save people some clicks since it's not possible at the
moment to switch the theme mid-game and see the results immediately.


Former-commit-id: 0fed7c206b57b49e32d3f4dd84f859bc736ba736
Former-commit-id: c602653131ad02ced51b13ef651c16516f870d56
Former-commit-id: 072ac2cc35cfa16ce48b15413ecfb0e9cc986866
Former-commit-id: 78a77231a4dd910d42e6daab775b177b82061fe3 [formerly e27166de7b0634283e4217652e54de2488a13ce6]
Former-commit-id: 99742c6cec8a36427cccb17c1cfba129432c7a13