theme ui: Implement [theme] id= attribute, defaults to name=
[theme] id specifies the internal theme id used by the engine to match
the theme definition against the recorded user settings. Existing themes
are still handled by using the name attribute if it's present when the
id is not, and a deprecation warning is issued in that case.

The name attribute is expected to be translatable for UI usage, whereas
the id attribute is not. (See also: [unit], [unit_type], etc.)


Former-commit-id: 40137bd0d5ddc8c841b818b150878ba328e81b7d
Former-commit-id: fd4223baa1ebe7bd02c57dc57f4ee206af6f53a4
Former-commit-id: a21a914e3f3757545334f19e249b90749be7720c
Former-commit-id: e7f7f5376810790022d5b5e8b39b55421f6bab98 [formerly c391405dbf57cb8614e443bea4af8a3f23e2fad4]
Former-commit-id: 7e8fd66dc0f2c2b0670fcc9d7a177d5d7bb7b168