HttT S17: Add a wmllint: noconvert directive amidst Lua code
wmllint doesn't know what Lua is or that it should skip it.

The code in question was triggering the viewing_side -> side plain text
substitution rule because wesnoth.find_path takes an optional last table
argument that includes a 'viewing_side' table entry, that is completely
unrelated to the [filter_vision] viewing_side -> SSF change in 1.11.x.


Former-commit-id: 3114349fcdf408807423fdd40da4920ef2b47b6a
Former-commit-id: 0b71ec5811d608469d87ac4447c045b9161e2f2c
Former-commit-id: 442a34d4b5ab00b2a797be0c5b97eed2b73e830c
Former-commit-id: 82a916e0d44297b1eeaa44181aa0131f5990cfd4 [formerly 531096178bfc08fa488ec1cc2e4cb2fa0a156a99]
Former-commit-id: cd624415512640b8dbcc988169b7544de64c9103