Rename some variables to reduce confusion.
The variable 'n' was declared within a scope that had already
declared 'n'. While each variable is only used in the line in which
it is defined (they are loop control variables for 'for' statements),
it's just bad style to have that potential confusion.

Plus, 'n' is not a descriptive name. (Not that the new names are
much better.)


Former-commit-id: 1e01ce11e0af1370e1d92e65a64463d436b7f3ab
Former-commit-id: 1d0457871466f1cbfb4a42a156bb115d04eb94df
Former-commit-id: 599ac895fc4f1919fca3b1dcaa36e0078097e99f
Former-commit-id: 8e331b6302ab24ce13f68bfecf7edd6ae7341cb5 [formerly 9829ca3b497147280ad5e643089765faca2d3a3a]
Former-commit-id: d305f34be0ad5e296ced15ff7c5b9c82ae4d7727