Turn dialog used for hotseat mp games now applies a blindfold
This is intended to address bug #3856.

Following discussion on IRC, this is considered a feature request
and not a bug, and so cannot be addressed during feature freeze
for 1.12. Hence I am pushing it to a topic branch.

I can't push a changelog change either as we don't know what version
this will go in, but the changelog entry should be something like:

"Fix bug #3856: The turn dialog used in hotseat MP play now applies
a blindfold to the map for the duration of the dialog."


Former-commit-id: 8022a0bdefc4f5f6ef7317c884832c8c8b79e9d7
Former-commit-id: 00e924e7ded5dcb8e6e85c0117f549883cc8b4b6
Former-commit-id: 8eb1d767f1b8616fc742f2ff40efac79c02f6360
Former-commit-id: ba04857bc6f74cf78a0dbda1d2417d9d32942b9e [formerly 8e023f9eb637079562779167c53d89ba5d97fde8]
Former-commit-id: 7c42b1ff4c6cfc3ef089ebf83a9c79f262959d98