Micro AIs: fix errors caused when WML events interfere with AI actions
If a WML event takes a unit off the map, many of the AIs would
previously have run into problems by trying to access a unit that is
then not there any more.  Thus, the existence of the unit needs to be
verified before continuing in cases when this would cause errors.


Former-commit-id: a8bb29ef0f8b552c6bbbb18339b2650dfc461819
Former-commit-id: 38d82d235c742ad86d43a052522828f421ecab37
Former-commit-id: 17533331adce8f45f10338e9b76bf60fb2abba1d
Former-commit-id: baf6e55769b5bfa89baff06f1f72b6989d28409f [formerly 72c6b1683e08f596494da595b5991eb06e3f80b2]
Former-commit-id: 775099745f5c595f6256ac041f2e3614274df0fc