gui2/tgame_cache_options: New Preferences subdialog for cache management
It allows the user to browse to the cache (using a desktop API call),
copy its path to clipboard, clean the cache contents (deleting stale
files), or purge it entirely.

This is ideal for people sticking to the same cache dir for very long
and thus keeping content that was generated by previous versions of
Wesnoth they might not even use anymore. Purging the cache might also
aid troubleshooting (although there is a --nocache switch for that too).

The dialog also makes Wesnoth slightly more transparent about its own
disk usage, information which may not be readily accessible to people
without the technical know-how to locate a hidden directory.

For now it's accessed by pushing a button in Preferences -> General,
even though this is not ideal since it is actually Advanced Preferences
material given the nature of the cache. Unfortunately, the Advanced
page's layout is currently monopolized by the advanced preferences
options and I'm not interested right now in solving that UI design
puzzle; it will have to wait, much like the Paths dialog has done and
continues to do for 1.12.


Former-commit-id: 495aee4c83a495d78fec01958d7a800f0fdc2866
Former-commit-id: 818459eb925d0e19826fa0c98dc4c207fd13bafb
Former-commit-id: e6d383692d5f6c725e2a25e00373562d177dadfc
Former-commit-id: e65e78131bcd7af986bc7ea1c10548ef6253369e [formerly cc21613dd2ef0258a7c6eafa98e9fc1cd76c6be3]
Former-commit-id: fd7b0b8f249d1316a0f767bf29497f917046d6e7