Make all labels visible in the editor.
I don't know why the earlier logic was so convoluted. Someone tried
to mash so many conditions in one line that apparently I missed the
intent earlier. I'm no longer trying to match the earlier logic, but
instead just trying to get the conditions right.

I really don't know how this will interact with the map editor's
label tool, because that tool is disabled for me and the screen
gives me no indication of how to enable it. But this does appear
to fix bug #21732 without messing up multiplayer.


Former-commit-id: a918bfaf6796aa660851e008336cb04ea27c4d57
Former-commit-id: dbc23e6af5be2d4f5084e58fa8e466670b54c218
Former-commit-id: 4650b56611f9dd5e8ccdd22d1942a750d6cae13e
Former-commit-id: cb983f639a80783bf291592d71fc6546a838a38c [formerly 16038c52c45f1918f7b6504e10b7f10375050534]
Former-commit-id: abdb2c0fcf5e2fe58a654cb75a174edd7ef1994e