Refactor show_video_mode_dialog.
Separated the dialog code from the resolution gathering code. This makes
it easier to port the code to SDL2.

This also changes the sorting of the resolutions, first width then
height instead of based on the area. Before a part of the list looked
like:
1280 x 720
1152 x 864
1280 x 960
1440 x 900
1282 x 1024
These now look like:
1152 x 864
1280 x 720
1280 x 960
1280 x 1024
1440 x 900
The new sorting makes more sense to me.

Former-commit-id: 82a393b718cf41ab9c54a14f7ca9bbb83fad0e3c
Former-commit-id: 124d641a242774694efdf3be2bc199c915dd4da2
Former-commit-id: 154cbffffcc9030a937e347009700ba174e35867
Former-commit-id: 80191c74df4977238f59d77b0f126f4644055efb [formerly 50e2a3a0f0098413bf324bc25eb588c2d37bbc5c]
Former-commit-id: f00f16143fd1b663a36974db5b7e3ebbd10399f4