Make sure cut_surface gets a neutral surface.
Also change its return to an optimised surface. Fixes bug #20876.

From the comment added to the source:
Unlike other image functions cut_surface does not convert the input
surface to a neutral surface, nor does it convert its return surface
to an optimised surface.

Since it seems to work for most cases, rather change this caller instead
of the function signature. (The issue was discovered in bug #20876).


Former-commit-id: a023cf5b03742929499d5f4f133c23649dcfca37
Former-commit-id: e7ae398ba452b475ffc428187c8a76b8e6a55ca7
Former-commit-id: 4f4960488f284fca9778d1cb424f005fa015309c
Former-commit-id: 763140cf221175cb17bd6defbf48bec79128d6f2 [formerly adec0f63033e764c1781d9c76db54ce6cb615109]
Former-commit-id: 57c0010baf3cf2aff8e6f32ea8e02cc72936f976