Make sure cut_surface gets a neutral surface.
Also change its return to an optimised surface. Fixes bug #20876.

From the comment added to the source:
Unlike other image functions cut_surface does not convert the input
surface to a neutral surface, nor does it convert its return surface
to an optimised surface.

Since it seems to work for most cases, rather change this caller instead
of the function signature. (The issue was discovered in bug #20876).


Former-commit-id: 5e915fe9d816c589cb712fe9907de63e647f8950
Former-commit-id: acd4f622ad77a2185c933fc94ce432cdf1223a24
Former-commit-id: 9c9830f9723b8527c4652e056678022530e31404
Former-commit-id: 35d8f7b70d2934b291c69215564ca18b169be23b [formerly 0a004dfeb715eacfcbad406130f8bd5fc19556b8]
Former-commit-id: 09a529579e641ee4e89c3f857947af137dbe7a92