Make sure cut_surface gets a neutral surface.
Also change its return to an optimised surface. Fixes bug #20876.

From the comment added to the source:
Unlike other image functions cut_surface does not convert the input
surface to a neutral surface, nor does it convert its return surface
to an optimised surface.

Since it seems to work for most cases, rather change this caller instead
of the function signature. (The issue was discovered in bug #20876).


Former-commit-id: b8915750e6c81c4e2309f2ae6f8fca1544fae964
Former-commit-id: 527a8fa74fb500e4d7bacde17dd807037b3d44ab
Former-commit-id: c113c11b7e61cda9c2ce48cbbbe386eb4950bb54
Former-commit-id: 7df80afbf1be0a66b37e7aa33676459cd9b499db [formerly bcf57db759eef18dc834365b5550bbd18c3fe4c2]
Former-commit-id: 8605a09c0d2eeea3096a5874b0d87ab4ab259a36