Move window caption initialisation code.
The code is moved from game_controller::init_language to
game_controller::init_video. This makes porting the code to SDL 2
easier. In SDL 2 the caption must be set after the main window exists.
It could be set in the constructor of the Window, but that can be done
once we drop SDL 1.2 support.


Former-commit-id: ecd4bda2a74ae21c1428f85a9ac2aa45ebd33ffb
Former-commit-id: f5e1c28f7a006e93efdbefc6d4b4a8d43823edaa
Former-commit-id: 844ef9a0e3aaa37d77b0a44d7e5b6474a1c866b0
Former-commit-id: 5335832493c66908dd4b4ca1d039169db759fe46 [formerly 24a20e26f8628fbb276de143c36535e1f1ef87bb]
Former-commit-id: 46eb6c85e62dd8b6d7cd70fba25cc7eac3912f78