gui2/ttransient_message: Hide title and image widgets when unused
When the title or image widget labels are left empty, the widgets
continue to take up space in the grid, which is particularly conspicuous
with our current border,border_size=all,5 convention.

Commit bf5a3a60c68adfb2487c3dd73f3b138bea0cae3c [formerly ef38c4d913bd422d070473906fc18d0f96c42039] [formerly 9ebc0c35133b4bd33e3edcf22a152567fd6cdd04 [formerly 2e37e7291dcb8e4633354af9ed67114e23d8f039]] [formerly ec389da66ca70e8efec0fd853c1ff0871f875403] [formerly d25ac27c6fcceedbf7a4811385265b205341e5a3] [formerly c542cd28b8a5196bb9ac9e19c14540c723a91c78] exposes the underlying
problem quite clearly in most dialogs because the image widget's cell
takes up extra space to the left of the dialog while empty -- see the
Objectives popup for an example.

Hiding the relevant widgets when missing label values is trivial to do,
even if it requires additional boilerplate in the
gui2::ttransient_message case (no existing pre_show() method override),
so it should be a good solution as any until cell borders gain the
ability to become null when the contained widget is otherwise unused.

NOTE: I had to move the inclusion of gettext.hpp around a bit in order
to avoid name collisions with Boost headers introduced along with the
two new GUI2 inclusions in transient_message.cpp. The _ gettext
shorthand macro is truly a PITA to have near some of the most
complicated Boost libraries.

Former-commit-id: 43f1c346a88c6265ae57cd7930a8b5675164dcec
Former-commit-id: 793f199b77a337fa35a7272e279bc80c763a07e7
Former-commit-id: b852e602287fb9dfb6ea89df5c59fb2121b720f9
Former-commit-id: 56acf674f4e4cb5c7d80fddc243b1a7c26b25954 [formerly 77021b1d1d17726dc5b01ef57c331a8629c0c682]
Former-commit-id: 381f96c340fb655aa0db715d2409903068374a33