gui2/tmessage: Hide title and image widgets when unused
When the title or image widget labels are left empty, the widgets
continue to take up space in the grid, which is particularly conspicuous
with our current border,border_size=all,5 convention.

Commit 2fce75a63064bd0f1a40313f4d41c7e11ebda1c3 [formerly 470282d1d4912ef41243704d37db46b31f98b94e] [formerly 8e6b394a0e34d536524345a2bf527c7c448474f7 [formerly 53992a3bc3407018b19907da2e1f057f58bd516e]] [formerly 84b2a576c96526d89ade6222b6ef3e2c185b0955] [formerly 4110b95278a8604f2a667c2f5091d0c9eab3d578] [formerly 0bad654cec7637cc8a9066c175e5f2126057cca4] exposes the underlying
problem quite clearly in most dialogs because the image widget's cell
takes up extra space to the left of the dialog while empty -- see the
"you have not started your turn yet, do you really want to end your
turn" prompt for an example.

Hiding the relevant widgets when missing label values is trivial to do.
In the gui2::tmessage case it required a minor internal refactoring to
keep a find_widget<>() reference around for a second method call for
each widget. This should be a good solution as any until cell borders
gain the ability to become null when the contained widget is otherwise
unused.


Former-commit-id: f566a603d115ec0115bc99d8d21908608d00262b
Former-commit-id: 1befc03f33cb6cdd2aecf2c0af77bf9770f99e89
Former-commit-id: 795d1beac93fea0fb1aa5a4baada2ba651933a80
Former-commit-id: 98d512127db9e6c2e131c00afa18d02f0e634edb [formerly 3d8637cf498213a26dcca0738b5bde96d19d758c]
Former-commit-id: 5b54a50ca2909498a78aa4afabd4925c0c2e0abf