gui2/tmessage: Hide title and image widgets when unused
When the title or image widget labels are left empty, the widgets
continue to take up space in the grid, which is particularly conspicuous
with our current border,border_size=all,5 convention.

Commit 2fce75a63064bd0f1a40313f4d41c7e11ebda1c3 [formerly 470282d1d4912ef41243704d37db46b31f98b94e] [formerly 8e6b394a0e34d536524345a2bf527c7c448474f7 [formerly 53992a3bc3407018b19907da2e1f057f58bd516e]] [formerly 84b2a576c96526d89ade6222b6ef3e2c185b0955] [formerly 4110b95278a8604f2a667c2f5091d0c9eab3d578] [formerly 0bad654cec7637cc8a9066c175e5f2126057cca4] exposes the underlying
problem quite clearly in most dialogs because the image widget's cell
takes up extra space to the left of the dialog while empty -- see the
"you have not started your turn yet, do you really want to end your
turn" prompt for an example.

Hiding the relevant widgets when missing label values is trivial to do.
In the gui2::tmessage case it required a minor internal refactoring to
keep a find_widget<>() reference around for a second method call for
each widget. This should be a good solution as any until cell borders
gain the ability to become null when the contained widget is otherwise
unused.


Former-commit-id: 90820d8bb421b1b67648e99959abf0c94e0617dd
Former-commit-id: 4ce55f677cbb2c3a522776bd902230c7a85da1ac
Former-commit-id: ac41ecdcbfdcfd5fc81b2038da06fbc5a58e5180
Former-commit-id: c2d269b3f69daac007c2158e88e68e109f932ce7 [formerly 7e00dbc931ec0029b16d9bbd0fb303f4142ff9fd]
Former-commit-id: 1d3c9d86479a6134d917ed31e967ce0cc91bf9bf