sound: Increase the mixer channel allocation from 16 to 32
This raises the sound effect (SOUND_FX) channel group size from 5 to
21.

The rationale for this change is that the current SFX channel group size
is not enough in some situations (particularly during cutscenes and with
accelerated animations enabled), and decreasing the channel group size
for sound sources (currently 8, preserved as-is in this commit) doesn't
seem like the best solution at this point considering that the sound
source feature is neglected enough as it is. Three other channels are
reserved for the turn bell, turn timer countdown sound, and UI sounds,
and it'd be an even worse idea to drop those.

The SDL_mixer documentation suggests that increased mixer channel
allocations may be a memory usage concern with a high number of channels
("We're talking REALLY high!", their words), but hopefully this small
increase is good enough for our and the UMC making community's purposes
without impacting memory usage or performance on our main supported
platforms. At least Wesnoth's RSS didn't seem to change over a megabyte
when testing this patch on Linux (using ALSA).

Note:

Should this increased limit cause issues on some platform that matters
(hopefully not!) an alternative would be to reengineer the sound channel
selection done in sound::play_sound_internal() so instead of skipping
sounds when all channels are busy, the oldest playing channel is stopped
and reset to the requested sound, for channels in the SOUND_FX group.
However, this seems too complicated to do at this point for 1.12, so yet
another alternative would be to simply reduce the number of channels
reserved for sound sources and give those back to SOUND_FX.


Former-commit-id: 68d4650a6c3d5810e9c4d583246dda22301dedc2
Former-commit-id: a01c759bb9b68ada6887af658480335a2b8fb74d
Former-commit-id: fc3df4e6d972cd03374319546120acd32b57831f
Former-commit-id: 276c2d3d5e2ff23d5dfdb77971b0f065701d43b6 [formerly f82939160619444c2497350a20bef05247c0e800]
Former-commit-id: 5aaf99f566e184aa2df10977affcea19f17e8c37