fix bug #18829
it turns out that when the "network_ai" controller type was introduced,
it was implemented on the server in the following way:

https://github.com/wesnoth/wesnoth/blob/40975de96a7ada14ee4191e05bb7874636626e51 [formerly 61a121b6f7642cba3d1f4bc58c8f2987abca76ec] [formerly 0889087ff25a4a1907f252e0a7a2b38efbec26cb [formerly 05712cb0fd37d362eaeb20662ae01a1c34651f17]] [formerly c3c5416d06e92a9242abe1cbc3d072e26a3513dd] [formerly d751ead957cb0a9bc2f56284c20729a24cbe057a] [formerly 11b37f5cd59af9d8cfe9a119063c88b033e5c2a0]/src/server/game.cpp#L464

change_controller requests to the type "ai" are always transformed
by the server into either "human_ai" for host or "network_ai" for client,
and in thunderstruck's refactor of the mp_connect dialogs, ai sides are
always set to "ai" in the WML sent to the server (and set this way on
the host).

however, because of code in play_campaign.cpp which long predated this
refactor, all sides on an mp client are either set to human (if that
player controls them) or "network", or "null".

this causes problems because if that player saves the game, the savegame
will record these sides at human controlled rather than ai controlled...
if the game is reloaded oos can occur, although I won't detail it here.

we update the play_campaign.cpp code to ensure that "ai" controller types
always resolve to human_ai or network_ai as appropriate. additionally,
we make sure that on the host, sides always resolve to human_ai rather
than ai. so the type "ai" should never be observed during an mp game,
and only during scenario configuration, and perhaps during replays.

additionally, we ensure in playturn.cpp that if the client gets a message
from the server to set a side to "ai" during the game, it will in fact
set it to "human_ai" or "network_ai", to preserve the invariant.

finally we also switch over observers to follow this behavior --
previously there was a hack here in the server which would make sure
that any observers which join see all sides as controlled by "human"

https://github.com/wesnoth/wesnoth/blob/40975de96a7ada14ee4191e05bb7874636626e51 [formerly 61a121b6f7642cba3d1f4bc58c8f2987abca76ec] [formerly 0889087ff25a4a1907f252e0a7a2b38efbec26cb [formerly 05712cb0fd37d362eaeb20662ae01a1c34651f17]] [formerly c3c5416d06e92a9242abe1cbc3d072e26a3513dd] [formerly d751ead957cb0a9bc2f56284c20729a24cbe057a] [formerly 11b37f5cd59af9d8cfe9a119063c88b033e5c2a0]/src/server/game.cpp#L197

in order that they do not attempt to substitute themselves for a player.

we change this by removing the hack, and programming the client to remember
if the player is an observer, and if so not attempting to substitute themself
at game start.


Former-commit-id: dcf3680d1ad3e06ba668432158ce60b50d744549
Former-commit-id: 05d21a3ccf2cb84c47e696b1a5278bc95e4ffb82
Former-commit-id: eaac9323f3cafd9dc76b93b6ed5366d7226e3ba5
Former-commit-id: 88b0a5fc782ac10b9b0ed26ede6e903fa53dacec [formerly 56cedf2c4a46eb794743970dbfa4219876bb6c62]
Former-commit-id: 54dcdad039068d4ad76926d586f6bb5d52359288