fix bug #18829
it turns out that when the "network_ai" controller type was introduced,
it was implemented on the server in the following way:

https://github.com/wesnoth/wesnoth/blob/40975de96a7ada14ee4191e05bb7874636626e51 [formerly 61a121b6f7642cba3d1f4bc58c8f2987abca76ec] [formerly 0889087ff25a4a1907f252e0a7a2b38efbec26cb [formerly 05712cb0fd37d362eaeb20662ae01a1c34651f17]] [formerly c3c5416d06e92a9242abe1cbc3d072e26a3513dd] [formerly d751ead957cb0a9bc2f56284c20729a24cbe057a] [formerly 11b37f5cd59af9d8cfe9a119063c88b033e5c2a0]/src/server/game.cpp#L464

change_controller requests to the type "ai" are always transformed
by the server into either "human_ai" for host or "network_ai" for client,
and in thunderstruck's refactor of the mp_connect dialogs, ai sides are
always set to "ai" in the WML sent to the server (and set this way on
the host).

however, because of code in play_campaign.cpp which long predated this
refactor, all sides on an mp client are either set to human (if that
player controls them) or "network", or "null".

this causes problems because if that player saves the game, the savegame
will record these sides at human controlled rather than ai controlled...
if the game is reloaded oos can occur, although I won't detail it here.

we update the play_campaign.cpp code to ensure that "ai" controller types
always resolve to human_ai or network_ai as appropriate. additionally,
we make sure that on the host, sides always resolve to human_ai rather
than ai. so the type "ai" should never be observed during an mp game,
and only during scenario configuration, and perhaps during replays.

additionally, we ensure in playturn.cpp that if the client gets a message
from the server to set a side to "ai" during the game, it will in fact
set it to "human_ai" or "network_ai", to preserve the invariant.

finally we also switch over observers to follow this behavior --
previously there was a hack here in the server which would make sure
that any observers which join see all sides as controlled by "human"

https://github.com/wesnoth/wesnoth/blob/40975de96a7ada14ee4191e05bb7874636626e51 [formerly 61a121b6f7642cba3d1f4bc58c8f2987abca76ec] [formerly 0889087ff25a4a1907f252e0a7a2b38efbec26cb [formerly 05712cb0fd37d362eaeb20662ae01a1c34651f17]] [formerly c3c5416d06e92a9242abe1cbc3d072e26a3513dd] [formerly d751ead957cb0a9bc2f56284c20729a24cbe057a] [formerly 11b37f5cd59af9d8cfe9a119063c88b033e5c2a0]/src/server/game.cpp#L197

in order that they do not attempt to substitute themselves for a player.

we change this by removing the hack, and programming the client to remember
if the player is an observer, and if so not attempting to substitute themself
at game start.


Former-commit-id: 404f56546d98135b3686ee312a6a43a1428276fd
Former-commit-id: 8537a002fed7bce6ee1eb0473c7794784df524a0
Former-commit-id: 9715fbaa7be2bda68d5c12b591b56edfdf5ec053
Former-commit-id: 454a598d1993078d819962b2e970e3a3c34830fa [formerly d451fc91bb20fb0626a8901d2e1c678087f7718b]
Former-commit-id: 30a95b6445c99460b1cf6700c3b043a9a4401adc