Implement unit instance recall costs.
Actually allows for the unit instance recall costs to be used. Proper
taking of the gold and undoing the cost for each unit is implemented
and works.  At advancement if the unit's current recall matches their
unit_type's recall_cost then it takes the new unit_type's recall_cost
otherwise it keeps whatever value it currently has until it fails this
check.  Recall costs are also now persistant and will carry over from
one scenario into the next as well.

If a cost is not set, it will continue to use the team/side/default
costs, these changes should not affect the current workings when
dealing with unaltered scenarios, and unit config files.  I repeat This
will not affect or change the currently way things work UNLESS a user
alters a file or includes the new recall_cost (as in copies and
modifies a unit in the scenario unit files) for a unit instance in
their scenario file(s).


Former-commit-id: 94a025c1975181580c97003846b7f21b12ed25eb
Former-commit-id: c80e90ab64e76870eaa47afe93a6e32cdfa7e756
Former-commit-id: b4d3840837f8ed1c26fe930ba10f75e423ad96b4
Former-commit-id: 4d3731ff6eed1db2ae04845d8da1730517cb5588 [formerly 6d7cd60ae696d3335dbe3723703c609e726e94e7]
Former-commit-id: dd743064107b17dbdbdbc256f44e73c97ca1ed8b