Show unit instance recall cost in Recall dialog
Show in the recall dialog below the unit type the recall cost of
that unit.  The costs are also colour coded, red for more then the
team/side recall_cost, white for same as team recall_cost and
finally green for below team recall_cost.  It also alters the not
enough gold message from "you must have X gold to recall a unit." to
"you must have X gold to recall this unit."  There is some new
comparisions to make this all work.  The header for one function
(recall_dialog) has had a const of the team.recall_cost passed to it to
allow for the comparisions more easy to do in that function.

There is room for improvement here, perhaps removing the costs display
if no unit has a recall cost that differs from the others.


Former-commit-id: 527eba0683c7af5ae4b296f2cbe16ea52152e77f
Former-commit-id: be38584721d8c0d1246335153d2556eb24d6a0fa
Former-commit-id: 02b6b8030220442e8805d7280453372a4cba1f28
Former-commit-id: 08e1feed1c8693bf8244ab6928d5402be92bae99 [formerly 8f3fd3002dfb72f0b7142ac760ba7649057230a2]
Former-commit-id: 54aec7cc3ed35f5524ac63866db19991b63b7bc8