Single unit instances work with whiteboard.
Here the changes to get whiteboard to work are mostly cosmetic, the
right cost for each unit is displayed on the unit in question.  This
is mostly for peace of mind as the user looks at the screen they don't
get confused.  The proper gold is also taken for each recall and
replaced for each canceled/undone recall while in whiteboard.


Former-commit-id: b95fa25dfc0c6e812f18a66f3a13e7ca5777ab1c
Former-commit-id: b0bff88b4188284c321d3ea817fb054ab6448bf7
Former-commit-id: b797392c39764eee696eff82131dc5d069a96810
Former-commit-id: 49785df8d9aa0df2b164abc378f0bc54f1e97bc4 [formerly 326d5afbf597cfbdc72355d9e7b0c9ecec3742c4]
Former-commit-id: b26d4fc3a8d78dbf2d934893d27e6793b74353d5