workaround for bug #21797
at a certain point in playcampaign.cpp after the level has ended, wesnoth was
grabbing .child("end_level_data"), and writing carry_over sides.
this would cause "mandatory child missing" errors in the test case described.

we fix the error by checking if the child was not present and if so adding it.
however, it may indicate a more fundamental bug involving reloaded games.


Former-commit-id: cd1570754ecfa83eb700d7a5fe2b537a4ce61020
Former-commit-id: 9daf2245bbd717c31bf9db1f4d615ea76138b9bc
Former-commit-id: 87c851c92f6159d8dc1df55addc6978b11825dc6
Former-commit-id: a0bf1f98d4daacb4a4cf3dd864f7d18e7674971e [formerly 759df216ef598ec2c2b63855820df7103628e255]
Former-commit-id: bfd6d05ede3b4dfcca79de85636e3aa027adf3db