workaround for bug #21797
at a certain point in playcampaign.cpp after the level has ended, wesnoth was
grabbing .child("end_level_data"), and writing carry_over sides.
this would cause "mandatory child missing" errors in the test case described.

we fix the error by checking if the child was not present and if so adding it.
however, it may indicate a more fundamental bug involving reloaded games.


Former-commit-id: 52d4ce29810bbcc86ad8b3d05b1685f76febca17
Former-commit-id: 1e6b747f4b092350e97f1a36820394cae9577006
Former-commit-id: 85052f2149ebd766f7033c5ac8f2bb7300e9003f
Former-commit-id: bbf80951385426a11e1813ecc7943cfd4fda1dfc [formerly 4076adb3f4ade645a2d6258f3867240c5b5e9057]
Former-commit-id: 849b9b8b63885cf5714a2b4a83722d6821b0d418