Fix for halos glitching through locations that become shrouded after rendering
Halos are never rendered for shrouded locations, but they are still
unrendered for them. A long-standing issue with [item] halos resulting
from this is that a halo that gets rendered to the screen at least once
for an unshrouded location later causes glitches if that location is
manually shrouded afterwards.

This can be avoided by not unrendering halos for shrouded locations. I
believe this solution makes sense because halos for shrouded locations
are never rendered first, and if the location is shrouded later on, the
display code will redraw the affected locations and overlay shroud on
top, causing the halo object's unrender buffer to become hopelessly
obsolete.


Former-commit-id: e0d354d968ae625c1dba8d9528dc27290b9f1c22
Former-commit-id: c5a74d3b9b074239ec4296073b3d418a824df064
Former-commit-id: 31ab590fa1a00697243f8dbc9f581396e20250d1
Former-commit-id: 35f39238f5f3bd456f36bf686e5f16637376d637 [formerly b0a215d56865777887b678a049d70628d58132b6]
Former-commit-id: 07e2a137d86885028ed8bd7ce3799a8de47ed1b9