Fix for halos glitching through locations that become shrouded after rendering
Halos are never rendered for shrouded locations, but they are still
unrendered for them. A long-standing issue with [item] halos resulting
from this is that a halo that gets rendered to the screen at least once
for an unshrouded location later causes glitches if that location is
manually shrouded afterwards.

This can be avoided by not unrendering halos for shrouded locations. I
believe this solution makes sense because halos for shrouded locations
are never rendered first, and if the location is shrouded later on, the
display code will redraw the affected locations and overlay shroud on
top, causing the halo object's unrender buffer to become hopelessly
obsolete.


Former-commit-id: 7c069fbb4751b00b4853ffef6a0bb40a06eda1fa
Former-commit-id: aaa439e6828b005bdac462502bbdca68fd7d71e5
Former-commit-id: 35eb9839ecc20dad3eda4e268201551684882950
Former-commit-id: 125eef5150f114b73346d758e6ad7c63ee02f2a5 [formerly c5a66d649c0a3d9b1fb3ffc7209173e4f5eda557]
Former-commit-id: 7f7c7acf47028d7a7ebfd1143d4a04568ba4b3c3