add synced_context class
the intention  is to fix #20871 and implement #21697, to sync
prestart/start events, and to implement
http://forums.wesnoth.org/viewtopic.php?f=6&t=39611 in the next patches,

the intention of synced_checkup is to replace random .. set/get_random_results.
Because set/get_random_results didn't have much to do with random, since it was only used to compare unit checksums and attacker damage to replays.

the intention of synced_commands is to move code out of do_replay_handle and make it callable from other places too so that we can just call the same function taht was called from replay in the simple cases (with synced_context::run_in_synced_command).

the object set_scontext_synced can be used to enter the synced context which enables synced_checkup and make the random calls synced.
Or we can use run_in_synced_context which is normaly easier than set_scontext_synced.
we can check wether we are in a synced context with get_syced_state to make addidional checks to detect oos (the other intention of that is to implement #21697 ).

this commit is part of pf 121.


Former-commit-id: a98388e5e7eb70bbc2d2f0c28f2d474372a3061f
Former-commit-id: 9dd3a086dd8d5ddf9e90714843f8c3cdcc1bf105
Former-commit-id: 12bf108b6e73b89b65c85dd8e42a4361433af5d9
Former-commit-id: 1633ec43b790d34105e268212d2cceb73d43acce [formerly 92adcdf8bd8f0036405bccebb91ea390dfdaf770]
Former-commit-id: a9f9904320ff8ec912fca804dc87fce5caedd82e