add synced_context class
the intention  is to fix #20871 and implement #21697, to sync
prestart/start events, and to implement
http://forums.wesnoth.org/viewtopic.php?f=6&t=39611 in the next patches,

the intention of synced_checkup is to replace random .. set/get_random_results.
Because set/get_random_results didn't have much to do with random, since it was only used to compare unit checksums and attacker damage to replays.

the intention of synced_commands is to move code out of do_replay_handle and make it callable from other places too so that we can just call the same function taht was called from replay in the simple cases (with synced_context::run_in_synced_command).

the object set_scontext_synced can be used to enter the synced context which enables synced_checkup and make the random calls synced.
Or we can use run_in_synced_context which is normaly easier than set_scontext_synced.
we can check wether we are in a synced context with get_syced_state to make addidional checks to detect oos (the other intention of that is to implement #21697 ).

this commit is part of pf 121.


Former-commit-id: fc7157e07b27743b548c90610bd3d7cf135e37fb
Former-commit-id: f5955e53402db009b98a4a776976f7196d1dd537
Former-commit-id: 17fe62a618b65244f307b550268ab4cabaa779ce
Former-commit-id: d059f8eb102f8b19ba2e86b6a881d2f8414b5e5a [formerly 4df9ca6ce1757e1ea7db85157e79436a28fcbd17]
Former-commit-id: 291893fc0eb9f3bb89ee96e9f6f8fd0f180c1cfa