make server level_ reflect true start of game position
Previousy, the [side]'s in level_ were updated with controller changes,
so that observers who join would be up to date, since controller changes
were not stored in history. We now change this, so that controller changes
are stored in history.

These are stored as "[record_change_controller]" tags rather than
"[change_controller]" because during a replay we should not
restart the turn and reinit in the event of such a change.

This is in preparation to move all controller tweaks to server-side.


Former-commit-id: 9559e31cc1c37473f26c47265bc06137d2444e33
Former-commit-id: e45b989df4a946aa8f8afee0a446b523b7d968d2
Former-commit-id: c589e6f6e535d1a6bdca0b52fc258d97f40a07b2
Former-commit-id: e8ca29484f52a62a2318c6065a9bb9dc43dc6e32 [formerly d43fe75a6830ad75ca5b9b2a5d543efb19c299dc]
Former-commit-id: ff2fa95a8336f2ab396b43fe5e9b088405bc6175