make server level_ reflect true start of game position
Previousy, the [side]'s in level_ were updated with controller changes,
so that observers who join would be up to date, since controller changes
were not stored in history. We now change this, so that controller changes
are stored in history.

These are stored as "[record_change_controller]" tags rather than
"[change_controller]" because during a replay we should not
restart the turn and reinit in the event of such a change.

This is in preparation to move all controller tweaks to server-side.


Former-commit-id: 0513a7639e9a639206172d17b180e0009c03a399
Former-commit-id: 347943fd7881a77d147e7be3e3b5e80c2341001c
Former-commit-id: ff20ce5026bc7d3d258aee6e22dc9c9b0e827bda
Former-commit-id: ea0a4ebcd76cd16143e58413015d01412b311d05 [formerly 3b586acc982a613172909725f35a668f271e972a]
Former-commit-id: 9b5477246ac3ed3bff9a542312124bc742406047