Single unit instances work with whiteboard.
Here the changes to get whiteboard to work are mostly cosmetic, the
right cost for each unit is displayed on the unit in question.  This
is mostly for peace of mind as the user looks at the screen they don't
get confused.  The proper gold is also taken for each recall and
replaced for each canceled/undone recall while in whiteboard.


Former-commit-id: 6f71217585691a0e4d52315dce09c3d5ca6d2d05
Former-commit-id: e072043167970bfe68690a6180d617548b72322e
Former-commit-id: 5b1f7c09c48b01814dd071b0761087dcf6b744d5
Former-commit-id: 9e19d60bf27fb7a62c71619b7db5a8d5cd70ec25 [formerly a02377cd2ee6357b1d99794e49849bf38eb84b57]
Former-commit-id: f0b9cdab2f8107071e98375b9ebc09e04e5956fc