Add missing MP settings for reloaded games. Fixes #21808.
The problam was that mp::configure is not used for reloaded games and so
mp_game_settings will not be fully initialised. This leads to using default
values rather than values set using mp::configure screen or from reloaded game
config.

The optimal solution would be to either display mp::configure for reloaded games
or to populate mp_game_settings using reloaded game config. However, both
solutions require rather a lot of code restructing and are probably not suitable
for a feature freeze.


Former-commit-id: 9dff943397cec4cca79ebe9ec8a8cfe708bfda08
Former-commit-id: d98b0a3143db3b221109f6764344b36372c1042c
Former-commit-id: 22c0fd435b8b7d76e5b6cb1e7d5a967d7e5073f3
Former-commit-id: 3efe1ddd2c67584e76898aca89f8ad6dc4ab657d [formerly eabc752a93248d73539d0688d89c9ce816eee0f6]
Former-commit-id: 645355a7ce7fd0940167123c8358c87eb0993322