Add missing MP settings for reloaded games. Fixes #21808.
The problam was that mp::configure is not used for reloaded games and so
mp_game_settings will not be fully initialised. This leads to using default
values rather than values set using mp::configure screen or from reloaded game
config.

The optimal solution would be to either display mp::configure for reloaded games
or to populate mp_game_settings using reloaded game config. However, both
solutions require rather a lot of code restructing and are probably not suitable
for a feature freeze.


Former-commit-id: c667a651bfcc959ec3b80cb2a7c2910dc5e15f73
Former-commit-id: df503ec81378a2c26c8e1bc690a048ebdbaf77ca
Former-commit-id: 5d4e6867880a1ba0f8f2c94c46898bca127ab6ca
Former-commit-id: e8f77a81bd568cab1265336f69172c9192eebea0 [formerly e5c29dc0b9f4c055359075bf4168e4c34e943bb8]
Former-commit-id: 04733aeab0bd2e159ee1b19397006a797cd28c89