Improve UTF-8 handling
For manipulation of UTF-8 encoded strings (insert, delete, truncate, and
size), usually conversion to wstring and back was necessary. This has
two disadvantages:
* the size of wstring is implementation defined
* the
double conversion is slow

The purpose of this patch is to add functions
to handle these tasks by a layer handling the different indexing between
codepoints and characters.

This commit applies these functions in
several places to replace the old implementations.

The byte_size_from_utf8_first() implementation will be replaced in a
separate commit as it relies on the count_leading_ones function by 8680.

Signed-off-by: aquileia <sk.aquileia@gmail.com>

Former-commit-id: 8099f7633a28df72bcca695261131ba370f4de46
Former-commit-id: 35976b24826ab06d85d32d3049444659015929a9
Former-commit-id: 71f6610a524586c8648a1437c3d2662f8974bcef
Former-commit-id: 7fbb7aac692b6a759546131afe506b61dc0c1a3f [formerly 4835c21906f701ecd512c5182cc0ddcb889f4a27]
Former-commit-id: d7f36dbdfe346db367bfa23f8f59a61aef9f6003