Improve UTF-8 handling
For manipulation of UTF-8 encoded strings (insert, delete, truncate, and
size), usually conversion to wstring and back was necessary. This has
two disadvantages:
* the size of wstring is implementation defined
* the
double conversion is slow

The purpose of this patch is to add functions
to handle these tasks by a layer handling the different indexing between
codepoints and characters.

This commit applies these functions in
several places to replace the old implementations.

The byte_size_from_utf8_first() implementation will be replaced in a
separate commit as it relies on the count_leading_ones function by 8680.

Signed-off-by: aquileia <sk.aquileia@gmail.com>

Former-commit-id: 63d1602f719ea513978e2063094f19b4cdc839e6
Former-commit-id: ebefb5683ee5912155c03fc48da1f32eef55033b
Former-commit-id: b597acf4a41718174d27771053c08a9cae96d7e6
Former-commit-id: 067af6def7547bccc6c7806d52b8a1a6a96210dd [formerly abbc24cd497b7d0800218b8b0212cfe7dee65f56]
Former-commit-id: c5debc67020d5410b728acbac0d69679ae7d4343