Fix bug #21759
Removed duplicate code that caused timer to be refreshed an additional
time if time runs out. The usual refreshing in
playmp_controller::after_human_turn is already called in this case as
well (after turn ends). The duplicate code gave the bonuses before the
turn ended, and after that the turn was ended (giving turn bonus again)
only if there was no random seed incoming from the server.

Conflicts:
	changelog

Former-commit-id: 0a80cd4f811f2ecd99894b42bad3d2006f7ad495
Former-commit-id: b206fef50d55b228e7c8d2d9869ff2b305fcd08d
Former-commit-id: 9ff11e56518f6c4e050f6a67c7ce78c290cc0be8
Former-commit-id: 36724a0b84c8220c74865ce584669e42b3d1727e [formerly e5050bf061481064eb1b71e31cdc2f4616f2e5c1]
Former-commit-id: ef936fb10e70b9b60e0c26cedfa82daa5e609731