Fix bug #21759
Removed duplicate code that caused timer to be refreshed an additional
time if time runs out. The usual refreshing in
playmp_controller::after_human_turn is already called in this case as
well (after turn ends). The duplicate code gave the bonuses before the
turn ended, and after that the turn was ended (giving turn bonus again)
only if there was no random seed incoming from the server.


Former-commit-id: cc8227392673d85645c80cac6e527243e73c08ea
Former-commit-id: 119ba4ffded73dff2ca7f6ed66676cd40a80062c
Former-commit-id: 30df205e197f3c1842e7d6ae6c086072349fc863
Former-commit-id: 255594c31363a271ef51c3006bc8a8e14cd8e807 [formerly 1354aa1ed0f40ce4e98ee1175067fea529f39ed7]
Former-commit-id: 385a9595aeac96eeec506bc74e32d3527137fcb3