Correct gold check comparision in recall cost
By correcting the assignment of the cost of the unit it now correctly
says whether or not the play actually has the gold to recall a unit, as
after this check it is assumed they do by the program which creates
the possibility of recalling when you have no gold, or negative gold.

Added comments so that the magic number is explained.


Former-commit-id: ead1e3ba2258a1dc1aa1d93ec3ce711449e09781
Former-commit-id: 7b50f53a2027637a5f0271d86e3de7592f299aa8
Former-commit-id: 84774bf6100c426057cc7d1badf29eca0360fb9b
Former-commit-id: 0d0e7c7189fa0db51c077fffe47e429b228b79bb [formerly a59b961bd95314b6ec4137d11f013895ce59a186]
Former-commit-id: 8de5ff7f1685ab363041871508d51a516789ee05