use synced_context: change get_user_input
previously and afterwards there can be 2 ways an action(attack, rectuit, move...) can be executed, the first way is that the action runs on the local client, and the action is sended over the network as soon as it is completed, this happends in recruits, moves, recall, and some more. The second way is that the action is sended over the network first and is then executed on all cielents simultaniously, this happends in attack events, and similar happends in prestart/start events. This is also why mp_sync didn work in this actions before: side 2 noticed that it needs data from side 1 but side 1 did send the execeution of the action before it generated the user_input data. That's why side 2 doesnt have the user input data at this point. Previously there was one special and very bugged case when the code waited for remote input before calling get_user_choice: [get_global_variable].

This commit moves (and fixes) the "waiting" from persist_var.cpp into get_user_choice so that the caller of get_user_choice don't have to worry about it. With this we can also use get_user_choice if we already sended data over the network. And we can enable the mp_sync in attack related events and prestart events. The intention of this commit is to fix fix http://gna.org/bugs/?20871.

the plan is to sync start and prestart events, so we don't need to check for that anymore,
we also don't need to pass the rng because we can use call random_new::generator->..

this commit is part of pr 121


Former-commit-id: 4f342f21f62b0d2d40ae30cc24a8a90169c58aa6
Former-commit-id: 79855cb224f34c75acda4855bf585863f0d2d3ef
Former-commit-id: a6da5ec749f02f8cad029e02e30adfc06e24d348
Former-commit-id: 35a09751d8feca57cd17af45487546acdd41ee38 [formerly 7ec7422b920ffb1336c64bddf593942fbdd476c0]
Former-commit-id: 7d09d34ba5b356930738601147d6b6f0aa81ad6c