use synced_context: change get_user_input
previously and afterwards there can be 2 ways an action(attack, rectuit, move...) can be executed, the first way is that the action runs on the local client, and the action is sended over the network as soon as it is completed, this happends in recruits, moves, recall, and some more. The second way is that the action is sended over the network first and is then executed on all cielents simultaniously, this happends in attack events, and similar happends in prestart/start events. This is also why mp_sync didn work in this actions before: side 2 noticed that it needs data from side 1 but side 1 did send the execeution of the action before it generated the user_input data. That's why side 2 doesnt have the user input data at this point. Previously there was one special and very bugged case when the code waited for remote input before calling get_user_choice: [get_global_variable].

This commit moves (and fixes) the "waiting" from persist_var.cpp into get_user_choice so that the caller of get_user_choice don't have to worry about it. With this we can also use get_user_choice if we already sended data over the network. And we can enable the mp_sync in attack related events and prestart events. The intention of this commit is to fix fix http://gna.org/bugs/?20871.

the plan is to sync start and prestart events, so we don't need to check for that anymore,
we also don't need to pass the rng because we can use call random_new::generator->..

this commit is part of pr 121


Former-commit-id: 5296cf2923329486baa63fa0df0eeb8ffb45450b
Former-commit-id: cfe29df57ead1f7a7933c9de869bdabe4c20ec49
Former-commit-id: 78c41ecfcfa3355db08d322e9324a0158b9f15f4
Former-commit-id: 5b5e19183a83f66838a2ec4738a6b2e6c426e023 [formerly 047a3b2bde21187db995a09895a60a8c5f8e341c]
Former-commit-id: bff4d704fcf1fe270dd07c15cfe2f49a48b94f98