use synced context: sync the turn ... events and prestart events.
by using set_scontext_synced and recording brefore doing that.

we split fire_prestart and fire_preload becasue prestart runs in a synced context and fire preload not.

we also dont need recorder.pre_replay anymore becasue of the changes to the rng.

the line synced_context::run_in_synced_context("auto_shroud", replay_helper::get_auto_shroud(true));
was accidently placed in this commit and should rather be in the commit called "use synced_context in undo and rest"
this commit is part of pr 121.


Former-commit-id: 340fa4e75f0e7b6c387195a4c70f31853f0671e0
Former-commit-id: 91d4600cfaf8741e12c9bf74ca5c012cc3b98d6a
Former-commit-id: 5f6e880e4f28979ddcc06f5c981e67dc1df73044
Former-commit-id: d402172b15e9f10b039b97fe55f45beadca5a540 [formerly 411aaecf80b0df37fd008f12c8ed02977e26b750]
Former-commit-id: 4cfdf4f25c45457b5417a2a7f08cb2284e20c493