use synced context: sync the turn ... events and prestart events.
by using set_scontext_synced and recording brefore doing that.

we split fire_prestart and fire_preload becasue prestart runs in a synced context and fire preload not.

we also dont need recorder.pre_replay anymore becasue of the changes to the rng.

the line synced_context::run_in_synced_context("auto_shroud", replay_helper::get_auto_shroud(true));
was accidently placed in this commit and should rather be in the commit called "use synced_context in undo and rest"
this commit is part of pr 121.


Former-commit-id: c707cc44366211e3d2ce1f870c6a44edd7e9f4e6
Former-commit-id: 17ba6849288126b42b5f708e465737997924914c
Former-commit-id: 9738341877899ff659b5c03ba692414df5a1f474
Former-commit-id: e6c84ce44e00e2e6f4b10c9b18b813ae20dde96b [formerly a23abb6018d4e7cfd192bad8ae6c68b55480317a]
Former-commit-id: fc72a2840b5f4f73d1a8d7d54c791ca94faa500c