add from_side check for require_random
"from_side" is an attribute whose only purpose is to enable a check by
the server that a package is really sended by the side given in
"from_side", if thats not the case, the server sets an attribute
"side_invalid". The main purpose of this contruct is to disallow
cheating in mp.


Former-commit-id: 8f1563ee0d95b5f42d3c4689a172a9eede5a9062
Former-commit-id: 3c0a973771cf52c0e9665fc5006ce278edab05b0
Former-commit-id: 978833a357a205878c3ca5540b4ee034804c18ea
Former-commit-id: 260e3148e8ee6cbae201b1ffa3de312d5cce37bd [formerly f1f31e0af2fccbfdafdb0b6471be3097861ea1f7]
Former-commit-id: 594e9c5278ed13f5052e9edfc6cdfb56deeea25f