add from_side check for require_random
"from_side" is an attribute whose only purpose is to enable a check by
the server that a package is really sended by the side given in
"from_side", if thats not the case, the server sets an attribute
"side_invalid". The main purpose of this contruct is to disallow
cheating in mp.


Former-commit-id: fddab6749d756becd5e67fe11f1c5297acb62e14
Former-commit-id: 9d9347b606e7787dfe9862db5f6b062d183ab193
Former-commit-id: 9b2a34788cac842023b3953b8ccff2b6777d90a6
Former-commit-id: dc6c25793cafbd41f607243cdf3ce0559e16a002 [formerly 5587fc4ca752165ab0ba17a92dfe195f805d1e80]
Former-commit-id: 6d7cf2e6465221cfe169cb6ba5a11c3460699a08