blindfold is removed when control of a side is received
in blindfold_ii, blindfold was taken off when local side takes control
in this fixup, blindfold is taken off as soon as host gives control of
any side.
to achieve this, playturn.cpp must be able to ask the playcontroller to
remove the blindfold.
we create a virtual method of play_controller, on_not_observer,
which is called whenever this client gets local control of a side
(it becomes ai or human)
on all play_controllers except play_mpcontroller, this does nothing
in play_mpcontroller it calls a "remove_blindfold" function
when a human side takes control, remove_blindfold is still called,
as a precaution.


Former-commit-id: b38838a0824eee8a67ed5d11bdb8187b3d5ae14d
Former-commit-id: 9de38d10f0ec7bb35000434102aca0781f865f24
Former-commit-id: 1b67b99020fb03905f434a879b305114524c3da2
Former-commit-id: 9160c7a22a74882729510467610de0a2ce852cc4 [formerly 819bb89fcce266ea1cdda37c94ccb3e48e67b683]
Former-commit-id: c339018beb462c94fea68db10aa610986a0b1d2c