blindfold is removed when control of a side is received
in blindfold_ii, blindfold was taken off when local side takes control
in this fixup, blindfold is taken off as soon as host gives control of
any side.
to achieve this, playturn.cpp must be able to ask the playcontroller to
remove the blindfold.
we create a virtual method of play_controller, on_not_observer,
which is called whenever this client gets local control of a side
(it becomes ai or human)
on all play_controllers except play_mpcontroller, this does nothing
in play_mpcontroller it calls a "remove_blindfold" function
when a human side takes control, remove_blindfold is still called,
as a precaution.


Former-commit-id: 54d563986dd901486c5adacde9db8ca36e145a74
Former-commit-id: 4c095ebb11971e60bfeb489b84f59e44a833ca99
Former-commit-id: 62737fccb749a81b158c5112f4988b335e1e6c9e
Former-commit-id: 408d5458e502c8e9b19e86247609c78eeb490f68 [formerly d4c9e0ed9351d9f705c1693b8d6bbd4133caaa37]
Former-commit-id: 0764a8a54d4364c652f9bb8147edea97515963d4