use synced_context in moves
In order to get the same move results that we got during the game in replay, we now save continue_move/skip_sighed in the replay.
We also call move_unit now with replay_dest = NULL, is_replay = false during replay.
So during replays we now execute the same code, that we execute during the normal game. That also means that, the variable replay_ in move.cpp is now always false. which means there is some unused code now in move.cpp
i still hesitate to remove this code because it somehow seems to be too easy to do it like this, but i have found no bugs.
i suppose, that is_replay_ if a leftover from a time when visibility wasn't calculated normally during replays.
in a later commit i also put skip_ally_sighted in the replay data.

I don't use run_in_synced_context for moves because some methods use the returnvalue of move_unit. Instead i splitted move_unit into move_unit_and_recod and move_unit_from_replay, because using 'if' isn't possible for set_scontext_synced. And i don't want to record a move if nothing happend. Also i think a very small code duplicate for one 'if' less is a very fair trade.

this commit is part of pr 121.


Former-commit-id: 9e54fe084e375adae3bcd5f42d76b7a5f5b18e21
Former-commit-id: ae54a2ec08ed2d9a0c2c87392f94fa14029600ef
Former-commit-id: de8d404afaea3613bea0fc64df80e7ed3e4e48f1
Former-commit-id: 9133415daab582f94b9c6567f1ba1acaa1260203 [formerly 0804046e8d7159ecbb61f3df8510a060712fb3dd]
Former-commit-id: ebd26bed216ce2e078552c3a38f5e0276f86b3fc