use synced_context in moves
In order to get the same move results that we got during the game in replay, we now save continue_move/skip_sighed in the replay.
We also call move_unit now with replay_dest = NULL, is_replay = false during replay.
So during replays we now execute the same code, that we execute during the normal game. That also means that, the variable replay_ in move.cpp is now always false. which means there is some unused code now in move.cpp
i still hesitate to remove this code because it somehow seems to be too easy to do it like this, but i have found no bugs.
i suppose, that is_replay_ if a leftover from a time when visibility wasn't calculated normally during replays.
in a later commit i also put skip_ally_sighted in the replay data.

I don't use run_in_synced_context for moves because some methods use the returnvalue of move_unit. Instead i splitted move_unit into move_unit_and_recod and move_unit_from_replay, because using 'if' isn't possible for set_scontext_synced. And i don't want to record a move if nothing happend. Also i think a very small code duplicate for one 'if' less is a very fair trade.

this commit is part of pr 121.


Former-commit-id: 052361c7e1354371c049915c0d749567aca9ec16
Former-commit-id: aeb9cb571c4a5da0434f90db38c12220a9884613
Former-commit-id: 7920e832da7ce0c2627cea542b246063274449b7
Former-commit-id: 80cb9167f8958aa19b5b1445fd2ce2e55f77aa6c [formerly 45cb0fec718203900136be9ed71b3adf79852bd2]
Former-commit-id: e19434403f6cce04838636dedde077a9f74bf084