use synced_context in recruits/recalls
we use ignore_error_function to get the same behavior as before. that means we allow to pass an invalid recall and just nothing happens in this case. We also don't put anything on the recorder, because run_in_sycned_context removes it from the recorder if we weren't successful, as intended.


Former-commit-id: e533385d181e072cae3227a91933277405ed26ab
Former-commit-id: 200634cdbe2c66f54a5024954d7e5983144f90f4
Former-commit-id: 571d54e1554f9fdc5c38d555f4cc325a2fccc7e9
Former-commit-id: 6130430c80d64ad2d4e3b391654ed66f82bd4cab [formerly 40f47348de8658f10a43d2a3d0bb2f7bdc5ee2fd]
Former-commit-id: 2e7b8a8231f445eb1d582676a3f4ce60fc711987