use synced_context in recruits/recalls
we use ignore_error_function to get the same behavior as before. that means we allow to pass an invalid recall and just nothing happens in this case. We also don't put anything on the recorder, because run_in_sycned_context removes it from the recorder if we weren't successful, as intended.


Former-commit-id: 0764094f3aeb4e90a707a9fa8551ff157010198a
Former-commit-id: f96b01e7d06fdac730df9756a465b7c8e471c9f8
Former-commit-id: dbd3ec97c7534d3dae55304d0f3903c05191c523
Former-commit-id: 4219cc9b0d2ed1550abce68063a4aa7889bcb193 [formerly ba0ebcc52e62700bf7e8000b1f866084133368c6]
Former-commit-id: 23d0b8843e0c508d28ca80dc93fbc095333e108b