add command line option --unit, -u, to run scenario as a unit test
This is the initial commit for --unit, -u, unit test option.
This option is a modified version of --test, -t. Scenarios will
run in the same context essentially, but will terminate on
scenario end.

The exit code of wesnoth is 0 (PASS) if the scenario ended with
LEVEL_RESULT VICTORY, or LEVEL_RESULT NONE.

The exit code of wesnoth is 1 (FAIL) otherwise.

Higher exit codes may reflect a failure for a reason other than
the scenario ending in defate.


Former-commit-id: 5f5515b1747147aabd50743042ff9ddf80a3b09f
Former-commit-id: 9fbaf8fd28d34ca51874eb338e773f3fcd329a3b
Former-commit-id: f38e50a72a71d4616c9db3b86f5d94e9fed16082
Former-commit-id: f6488409b07be09eae6df1f19a48e0ee3ed05679 [formerly f6adce9da2deada2b14bf6c653a193ae99305a76]
Former-commit-id: 12778ebe5ebe0d340f0f9f4c9e4d1cd86bdcf9a3