add wml_unit_tests test aggregator script
This bash script runs a sequence of WML unit test scenarios.
If all pass the result is 0.
If any fails, the result is 1, and the std::cerr stored in "error.log".
We also add error.log to the .gitignore list.

It shows how --timeout and --validcache may be used in such a script


Former-commit-id: 86211065f1bf523190dea41a5729d25461dfb50b
Former-commit-id: 91495140d5d97496d3264aa117e13bf9ad92c3de
Former-commit-id: a0022cd6f5f574c7c26eb3e4949cdd3f797b22f8
Former-commit-id: e2df3c318abd74b3184e4dcab95d057545a0b068 [formerly 695bbe33f5c126d42e5a43a892f33d0c1d7617d9]
Former-commit-id: 16c8c52939d62152aad8dcf2ae349182fca81269