Micro AIs: use AI id instead of CA id as identifier
… for determining whether AI/CA ids are unique.  Also pass AI id to
eval/exec functions, instead of CA id.  This is also a step toward
fixing the bug with MAI variables remaining stored in units after an
MAI is removed.
In principle, the ca_id= key in the [micro_ai] tag should now be
renamed to ai_id, but that would break backward compatibility without
any benefit to the user, so we deal with it internally behind the
scenes instead.


Former-commit-id: dac4a148079f7780694c9ca54c8c96123eafacd3
Former-commit-id: 01d05f2d20c42ed9a9316f6ab3dd643c6c9df53f
Former-commit-id: 6bcfb9ce0d10ef8ab2ab8fe9b401eac9cfab44cf
Former-commit-id: 7c02cbeacd34e1cf199bb2f823a82b48b610d6c8 [formerly 03d2b3bbc820a493c1445f336c462cd9d8f3ca09]
Former-commit-id: 8bd10c9bf3985420d775352139f97195fe260c3b