Micro AIs: use AI id instead of CA id as identifier
… for determining whether AI/CA ids are unique.  Also pass AI id to
eval/exec functions, instead of CA id.  This is also a step toward
fixing the bug with MAI variables remaining stored in units after an
MAI is removed.
In principle, the ca_id= key in the [micro_ai] tag should now be
renamed to ai_id, but that would break backward compatibility without
any benefit to the user, so we deal with it internally behind the
scenes instead.


Former-commit-id: 74026a2b10f15029613dc0445ce0035a58c89071
Former-commit-id: fa5b5a220155b7c268364caa684ca40af0cbf907
Former-commit-id: 81d474236ef3e29181a5dd7a3b254eea57158f75
Former-commit-id: 1b1d352487a0d540e58d67e0294dd19deb792380 [formerly 01b9b093bb4a5c1d6151bc2b8656dd39d6799255]
Former-commit-id: d11978dc77ce05642c7343782630beb470e8c1ba