implement the deterministic mode for sp. part2
the player can now choose whether he want to play a new SP game in the
deterministic mode where he'll get the same results when he reloads a
game.

i think it would also be nice to enable the deterministic mode for MP for players with verybad network connection, because that would result in less network traffic but more options to cheat.
but i currently don't know enough about the mp connect code to that.


Former-commit-id: 9111bbe7f98dc02a4097cdfd95ac2cbcdbcdaf91
Former-commit-id: 363bb9679a57b96bca4ef0d6994b343fb76db006
Former-commit-id: dff6674e990d52f862fec88d24e2cca0bea43573
Former-commit-id: 85fcc057df8186ae3e8ea122229f9cbdf553f435 [formerly bfd5b355fc39003a1bb60fefe590e0b636430364]
Former-commit-id: b290f365849474c7cab645a9a972815599e36c85