Upgrade Xcode project to enable both i386 and x86_64 builds
The largest part of this consist of upgrading the boost and pango/cairo
libraries together with their dependencies, supporting files and
project settings. In addition, architecture, SDK and compiler options
are set to use the native versions of the given OS X version for debug
builds. The OS X deployment target is set to OS X 10.8, because of a
change in the default libraries used by OS X from 10.8 to 10.9.
For this to work, the resources in  wesnoth_compile_mac_1.13.zip need
to be installed as well.
Some other settings, e.g. font defaults, still need to be adjusted.
This will be done in separate commits. Settings for release builds will
be changed later as well.


Former-commit-id: 894c502484478cf58a6937d6a90c7f8255aaea42
Former-commit-id: 056463022a6c1c1963ca98026578ec4c350a60e8
Former-commit-id: dc4475103bd0609c72eb95e6425b9b3e424af159
Former-commit-id: 2a0886ae7e2eba6938ec17d495229f573d0bd8b0 [formerly addd5a55b108c60f674f7b8d4546d5db86a51aaa]
Former-commit-id: 67bcb5fe0f7bfd3c0574ec152b64306ce24d3eb0