enable undo recruits/recalls if no rng was used
If we used rng during an acion we can never undo it becasue it is an nonundoable information gain. Also it's technicaly impossible because in mp using rng will result in data send over the network, which cannot be undone. Wedon't need to check whether the unit had traits/names this was becasue that would also have caused rng calls.

Former-commit-id: 9e7309b655c6cac65bc3023d54717763aa0014fb
Former-commit-id: befb375246bd118021521312a491463d91d4aa37
Former-commit-id: 93c5e38e0107dc646d0289c605592258707448a7
Former-commit-id: 4760c7737701d279bbb9cbdf1e63803d1d2acea8 [formerly 5786574417fe7493b0bae84f950afd084526bd04]
Former-commit-id: d896d86a0beabb00520460cc8165095392300be5