enable undo recruits/recalls if no rng was used
If we used rng during an acion we can never undo it becasue it is an nonundoable information gain. Also it's technicaly impossible because in mp using rng will result in data send over the network, which cannot be undone. Wedon't need to check whether the unit had traits/names this was becasue that would also have caused rng calls.


Former-commit-id: 3b1a428f80d3bf7d68d5d6ec62148bc182511cab
Former-commit-id: ab7b26288bf78ac7a1bda72669493ca37991aed1
Former-commit-id: e8cc062d1fa2c5ddfb3ca48b3641adb22e4b53ca
Former-commit-id: 935c6963b65cddfbf8924c6acf19b27aefb73e45 [formerly ac2597ab5fbeb8d215a233e7db7195b4c612a50a]
Former-commit-id: c5ead193ed8c7905d39bcf65eb76a624bbb740bc