fixup party full bell
the party full bell code is moved back to the engine, and is
based on checking whether processing the most recent message
has made it now possible to start the game. this seems to be
the simplest logic that will do the right thing, it will now
make the bell play whenever the "I'm ready" button is active

this also addresses bug #21910.

note: the reason we don't check the "silent" flag is that it
causes the sound not to be played in some circumstances where
it should. multiplayer_connect_engine currently returns
value "true" for silent (the second part of result) in all
cases except when a player has been successfully assigned to
a side, in which case it is false. depending on the status of
allow_changes, that might be a time that we want to play the
"ready to start" bell.

Former-commit-id: 82e33ff156920e54a63f27b5a95fc7b0a6368c60
Former-commit-id: 3c70fbf1bb9a9f116a2fb3c736ba6ca9e180bebe
Former-commit-id: d036632466558795ee491ef49a3794faea9bd325
Former-commit-id: 86b5fec47159943ccee9679c9230572851c9d3d0 [formerly f7c8aa4fc8272151ab769f51969a074af21e3ea9]
Former-commit-id: 3e2e9845e78be8c61b60dc7284360b1e426824a1