fixup party full bell
the party full bell code is moved back to the engine, and is
based on checking whether processing the most recent message
has made it now possible to start the game. this seems to be
the simplest logic that will do the right thing, it will now
make the bell play whenever the "I'm ready" button is active

this also addresses bug #21910.

note: the reason we don't check the "silent" flag is that it
causes the sound not to be played in some circumstances where
it should. multiplayer_connect_engine currently returns
value "true" for silent (the second part of result) in all
cases except when a player has been successfully assigned to
a side, in which case it is false. depending on the status of
allow_changes, that might be a time that we want to play the
"ready to start" bell.

Former-commit-id: 2e66bfa2566a7dcc5608363599a25919646da7c9
Former-commit-id: f5ffc3d3634b6ae92ade18088bcac332223dd732
Former-commit-id: b032aafe37df38a67e98e7c10a629e3679987521
Former-commit-id: db2ed1a5d60bcb5ea30968dfe7267013597bf53d [formerly 667c1aa7beea6034b77cd3ab7f492d338308987d]
Former-commit-id: 8086f0ec378c3eac08163f309158a5deb3a8c6fb