a check for deterministic mode in mp.
we now check during the start event whether the game is a mp game which
runs in deteministic mode and give an error in this case. This case
shouldn't cause OOS but it allows players  to cheat. I don't do it
earlier because i cannot use the screen during prestart events, and i
don't know about the code earlier (mp_connect).


Former-commit-id: fe59ac38a13b722f1f92a957fa8e2d2394a48da4
Former-commit-id: 397344e67e0efd0028405736f897b7e48b5ee846
Former-commit-id: adf58481cfea98a9ce760e097acfc0800913e916
Former-commit-id: 72c9fb17cacf23710b54ae5eff06a722e6e1bba5 [formerly 0ae0081339ca1bccffe33bd589c6001943fc9082]
Former-commit-id: 015c521d6078f111f57393848852d27f00aa72b0