provide a desktop notification when game is ready to start
this is played at the same time as the turn bell, it is effectively
the same as a chat message which will cause a desktop popup on
systems that we support this for.

note: we don't check notification preference, because this should
be done in the show_notification function, not the caller.
however, it seems that is only done with DBUS notifications? why
not others?

Former-commit-id: e3471e7702e9dda34d9c8243f8d5d7920b611144
Former-commit-id: 93f0ec641524a627c5d160f1ff04c6f0f9666843
Former-commit-id: 810c87e16ba9f099d1b62f67b08349327e291411
Former-commit-id: bbd1bab01860f54c98513092f4da46a9ad9cd318 [formerly 27b2de268d79b514d1990575c2a08a2c55058b04]
Former-commit-id: dd7289a056517a554474ce0b739b4c2306dc18e2