fix host reassignes side during choice
Assume the following situation:
2 Players in the game (A (side 1), B (side 2)), and one Observer (C).
it's A's turn.
A has to do a decision (for example unit advanement), A leaves the game
and B decides to assign side 1 to C.
On Client B: When returning from process_network_data side 1 is now controlled by B,
and then gets then reassigned with the next call of
process_network_data, but that's too late becasue handle_generaic_event
already returned and B does the advancement decision for side 1.
Players C gets the the first [change_controller] (A -> B) and then gets
the second change cntroller (B -> C) before getting B's answer to the
user choice. So when C gets the second [change_controller] he does the
decision himself becasue its C's decision an it has not been received
yet. -> the same decision has been made twice -> OOS.
this fixes this problem, by making the host wait for the servers change
controller in this case


Former-commit-id: 042fe6c6937ab85323f45cde093fc95b0236c081
Former-commit-id: ebda31817c102ee7d1be8f796a5d4193aa6e34da
Former-commit-id: 652173359b2110fbd62d090d999928a6078b6df8
Former-commit-id: 86c0c58de0e0c6579a054772ffa8dc2c442efd10 [formerly bfe43f8873e2e76c37b6dd0e512d3656c22f9fa8]
Former-commit-id: 7ea6cb42d89eddbfc6de217c8a7c75e64657e03a