fix host reassignes side during choice
Assume the following situation:
2 Players in the game (A (side 1), B (side 2)), and one Observer (C).
it's A's turn.
A has to do a decision (for example unit advanement), A leaves the game
and B decides to assign side 1 to C.
On Client B: When returning from process_network_data side 1 is now controlled by B,
and then gets then reassigned with the next call of
process_network_data, but that's too late becasue handle_generaic_event
already returned and B does the advancement decision for side 1.
Players C gets the the first [change_controller] (A -> B) and then gets
the second change cntroller (B -> C) before getting B's answer to the
user choice. So when C gets the second [change_controller] he does the
decision himself becasue its C's decision an it has not been received
yet. -> the same decision has been made twice -> OOS.
this fixes this problem, by making the host wait for the servers change
controller in this case


Former-commit-id: e315bf86fd92bd646867d12ac8a780c3947bc154
Former-commit-id: 5e141bd74825e8dddebfb7bcd9d79af7c5b7a544
Former-commit-id: 6e6b13be82448cad86d322dcc3e83ef6a8a349d1
Former-commit-id: 0c8e640a901596c224e05d412f06ca3b83751bb9 [formerly a33b381c01bffe61c68c6745c903870f092f9b4e]
Former-commit-id: 12d24334653b07c9e0a60cc158bc39edecd7415c