fix bug #21916, ready blocked at inappropriate times
the first problem was that the "wait" object and "connect" objects
didn't agree about when the joining player should be choosing their
faction. we follow the mp::wait object's convention, that it happens
exactly when "allow_changes = yes".

the second problem was that params.saved_game was checked by
mp::connect_engine at the wrong time. (specifically the side_engine)

it was checked at the time of making a new config for the side, so
that the config sent to the joining clients would have
allow_changes = no in a save game, but not in the constructor of the
side engine, which would have the default/input value regardless of
the saved game status. the result is that the host and clients are
out of sync about whether allow_changes is true for a given side.


Former-commit-id: b30c5e3d488acc44f2d2e49707470c026801dcb9
Former-commit-id: 021848eab85efb2667203f5e906e8c238fa0c9a8
Former-commit-id: a0d66409f4c9b358423c6361da39453f3853ac42
Former-commit-id: 29a02733c40d3dcee18dd771c1b7a5e2972b2b99 [formerly 259ec857278cba47b931f0503fbdee817d11daba]
Former-commit-id: c0a103c2e590c19273c432dc07125681e4df2cb9