fix bug #21916, ready blocked at inappropriate times
the first problem was that the "wait" object and "connect" objects
didn't agree about when the joining player should be choosing their
faction. we follow the mp::wait object's convention, that it happens
exactly when "allow_changes = yes".

the second problem was that params.saved_game was checked by
mp::connect_engine at the wrong time. (specifically the side_engine)

it was checked at the time of making a new config for the side, so
that the config sent to the joining clients would have
allow_changes = no in a save game, but not in the constructor of the
side engine, which would have the default/input value regardless of
the saved game status. the result is that the host and clients are
out of sync about whether allow_changes is true for a given side.


Former-commit-id: 8a582c7a76195fbf932c274524a14afb31e261d1
Former-commit-id: 858dd751368c499c5b464a0b4565f9c67148b142
Former-commit-id: c94699b15659f5f5e7dac7359e92b9bf6bf912b7
Former-commit-id: 58f5a5dba66f7358cc0502aecc81e296513e59cc [formerly 4a61886a3d0b3841c2b7ad1e57d1f43265aea625]
Former-commit-id: 675c54fbedbdcd20df91b7c87adb83b131209648