Fixes for factions, leaders and genders in networked games.
There was some oversight when changes to factions, leaders and genders set
up were made and as a result, in some cases, things were seriously broken.
This is an attempt to fix existing major problems. Some minor problems still
need to be addressed.

Specifically, this commit should fix problems with showing different factions,
leaders and genders lists for non-host and host players for the same sides. And
it should also avoid a situation where non-host player sees a "Custom" faction
even if it should not appear.


Former-commit-id: 6c3310f0d8713746fc3168ae99ddba965968b805
Former-commit-id: c6e54544acb908b86eeb2ddf63281eb60ddc2c70
Former-commit-id: 469009f3acc47dfefb0a3430c45ee97443004950
Former-commit-id: e19817a244c3f45e77370bd0ec680e6c23952f8f [formerly 84e0451290b93d4210004fc154a83369b43157d2]
Former-commit-id: f9cc842135a280224c9c51b2122366d8aa22b09f