Fixes for factions, leaders and genders in networked games.
There was some oversight when changes to factions, leaders and genders set
up were made and as a result, in some cases, things were seriously broken.
This is an attempt to fix existing major problems. Some minor problems still
need to be addressed.

Specifically, this commit should fix problems with showing different factions,
leaders and genders lists for non-host and host players for the same sides. And
it should also avoid a situation where non-host player sees a "Custom" faction
even if it should not appear.


Former-commit-id: 37a59256051664a63947f03849ca9c707e20bd23
Former-commit-id: 8c00cc1dad76328aa4c8ef6a774dba93b27e9926
Former-commit-id: 40fc0762e9731f24aeef8ae0313be193edfc7efc
Former-commit-id: 79ac3cdcb254b1d0721c2811c65cf224ed2f4f12 [formerly d4d326cde085e073aba1501f459d2b1d2e2559b3]
Former-commit-id: 9e371a59171db9b721a079b81a4294f8a7e892d8