Micro AIs: prevent potential conflicts of stored data
It’s theoretically possible that different Micro AIs’ evaluation
functions store information in the same variable within self.data.
This would only happen if the CAs have the same score and while this
should generally be avoided when setting up a scenario, it is better to
ensure that it cannot cause conflicts (not in the stored data at least,
the MAIs might still interfere with each other in other respects).


Former-commit-id: 163f9543157a2bb433635a2a8812ebfb06fc56c4
Former-commit-id: cd5c3c74e6fab415d835f3f1740afc7cf68a8dde
Former-commit-id: 07d63f4d00c5a652389c2da258f420eace8c2bea
Former-commit-id: a43e7c9b3c17a1836c13ad43bf84f68ce0a748b0 [formerly 3cd0e846c2644b0656de165b43f34a180cb547e4]
Former-commit-id: 8a912c28363762231dabe34ca33bf642dc9ddbe3