Micro AIs: prevent potential conflicts of stored data
It’s theoretically possible that different Micro AIs’ evaluation
functions store information in the same variable within self.data.
This would only happen if the CAs have the same score and while this
should generally be avoided when setting up a scenario, it is better to
ensure that it cannot cause conflicts (not in the stored data at least,
the MAIs might still interfere with each other in other respects).


Former-commit-id: 087bfaa111d543dac328c34dcecbea7c1a5b8983
Former-commit-id: 8884583913b77c0136675743f058985d41d63519
Former-commit-id: 14288c2e085ce94593462f8bf6c0a517be4c033a
Former-commit-id: 83bc7d3c2165c9edb3b4c5e94563b409c3200d8e [formerly e38a12ec9ee375ed9cdadb29eb5768b569e7fb20]
Former-commit-id: 9e3475ca535b86a7d2a601993d3e839c78376bb0