wmllint: change "usage of" magic comment to re.search
There seems no particular reason to require that this magic comment be at
the very beginning of the line, so why not switch from re.match to
re.search.

Also, update comments to reflect the fact that UtBS no longer uses this
magic comment.

A suggestion has been made to get rid of this magic comment now that UtBS
is no longer using it, but it may still be in use in some UMC somewhere.


Former-commit-id: 04f244283a154db22ccfdb76c067b98696c57026
Former-commit-id: d9f23f40c12a6527a54d62cde466ccfa41593320
Former-commit-id: 80c7472abe297798124a2bf6a64f8c55f3b2389b
Former-commit-id: 170a98e8d110fbac2bdb4f7ecd71bf7d0e0bc371 [formerly 1c9211fd63ec57e06d45e123d25db1db99e6572f]
Former-commit-id: 7d1b8870f0b5c9241dabbd9d3079d32dfe10dd94