wmllint: change "usage of" magic comment to re.search
There seems no particular reason to require that this magic comment be at
the very beginning of the line, so why not switch from re.match to
re.search.

Also, update comments to reflect the fact that UtBS no longer uses this
magic comment.

A suggestion has been made to get rid of this magic comment now that UtBS
is no longer using it, but it may still be in use in some UMC somewhere.


Former-commit-id: 63891c2aadf283e8ff68d1611dd6bbdf9591792e
Former-commit-id: 701316fb87bbaae79b56d06e47e77349a5981c01
Former-commit-id: 2433de13a9bf984737a5e3406a103e0c966ca769
Former-commit-id: 34cadf7a62f84f52b53ae5b1d471eada43a8a289 [formerly 277ff71d8919d28479043840e2b143b5ace08517]
Former-commit-id: f9b1acfc99ffd657db6db183b5c181b09b885a72