add [start] tag to server replays as needed by gfgtdf
this should have been part of 121, but wasn't
digest from irc logs:
post PR 121, all [replay]'s are expected to begin with an item
[command]
    [start]
    [/start]
[/command]
which may also contain some [checkup] children which help to
ensure that we didnt go OOS in prestart.
these commands are not generated by the server and distributed,
because we want the clients to begin loading the game without
waiting for the server to finish.
all clients, including observers, will add them automatically.
the server therefore should not add them to the history, and
rather just prepend them to [replay] when generating the
savefile.


Former-commit-id: 8616e03a2270005396c4de545e458bedc7e08a4e
Former-commit-id: 3d884c4051e556afe4fede88fc91243452e57123
Former-commit-id: c410c683244df8011a9a3e8ba58ab671012c9882
Former-commit-id: d98fb422b412bc3bce2e05dd6a7f5dedbe67066c [formerly 81092f1b59d8ad67f2a19c5fd448079c4cf5d33c]
Former-commit-id: a03239ffd3b3135792b0aaa90a8082e74a7d941a