add [start] tag to server replays as needed by gfgtdf
this should have been part of 121, but wasn't
digest from irc logs:
post PR 121, all [replay]'s are expected to begin with an item
[command]
    [start]
    [/start]
[/command]
which may also contain some [checkup] children which help to
ensure that we didnt go OOS in prestart.
these commands are not generated by the server and distributed,
because we want the clients to begin loading the game without
waiting for the server to finish.
all clients, including observers, will add them automatically.
the server therefore should not add them to the history, and
rather just prepend them to [replay] when generating the
savefile.


Former-commit-id: 86a5e163eb3bc087a5ca525105da2087f9194d5a
Former-commit-id: d77f1ec31c5c1986aad0bed7aa2bb44836d6471b
Former-commit-id: 318444cf5ce38cf55ca04bffca9cbfaee3bfd3db
Former-commit-id: ff20c24e7dc604eb07f8c769f45dd3b676f22c26 [formerly 7266623882816b77ac973280a41131b2ec7dc0ba]
Former-commit-id: 8afe20495a2dcdbec93178091ef505278bdcb562