always allow SAVE_REPLAY
We diallow save game during action because that would lead to broken saves, but this is not true for save_replay.
If, for some reason, we get stuck inside a an action we cannot save the game, with this we can at least save the replay in this case and create a vaild save with the 'replaying + saving durign replay' strategy.


Former-commit-id: 4d8ce244aa1e993caf87e3ddd2e9736aa83b41a6
Former-commit-id: 8f5469921349fc3c9e55ba5e25da525823addd0c
Former-commit-id: 06f75b0f40a1b00643fc95f0ab116db5828c6dc1
Former-commit-id: 330c05591a272545f8e5e82a0fa545d2cfd010b6 [formerly fdc489d321350070c577732489406e11b0549e77]
Former-commit-id: 0f9c1d3d1635e3cfdadc3a1d8fc931d4022eda67