always allow SAVE_REPLAY
We diallow save game during action because that would lead to broken saves, but this is not true for save_replay.
If, for some reason, we get stuck inside a an action we cannot save the game, with this we can at least save the replay in this case and create a vaild save with the 'replaying + saving durign replay' strategy.


Former-commit-id: f2ffa0516827823adcc02a436302f17f81a012a9
Former-commit-id: 3233ee65a2fcec0045b9d29859724964074d3f4f
Former-commit-id: a1cdc7b9198814dcc5a972ed22c97ddc9d632ea4
Former-commit-id: 645b9820ddd8ed2e40197f5368a55c58a0d1f805 [formerly 2b90b2e2c43797821f17d296d65c8753fce59973]
Former-commit-id: 3f26d6c12bc73770497d3f7ae00e6561b71f9f1f