LoW: Initialize units in start events
previously this was done in side 1 turn 1 events because start events
wasn't synced.

After the fix for 21933 players now loose when they have no units after
start events, so we need to do this in a start event.

replacing check_victory with check_end_level after start events in
play_controller.cpp line 572 doesn't work because we still have the
check_victory in playmp_controller when waiting
for remote packages, that check_victory is needed in order to sync the
behavior of ai turns which calls check_victory after every action on the
local client.


Former-commit-id: 467d648b31a31972d0ab21ca55296c9164566ae9
Former-commit-id: b61f1266b9fd5642f11bc2e06711a24648991816
Former-commit-id: 00b4f663c456b304b4058498b201f266981569e8
Former-commit-id: fa0d91949e6023da71fc1681859833677353fc33 [formerly da975c7a3439e9d6ded081f02894ae21a2ae8ef6]
Former-commit-id: 66eeb1700f25e9ea7f6d6bb444bce5a99800e80a